### 2021-06-11 #3359 @whedon
[PRE REVIEW]: grapesAgri1-Collection of shiny applications for data analysis in Agriculture-Part 1
**Submitting author:** @pratheesh3780 (<a href="http://orcid.org/0000-0003-3290-0436">Pratheesh P Gopinath</a>)
**Repository:** <a href="https://github.com/pratheesh3780/grapesAgri1" target ="_blank">https://github.com/pratheesh3780/grapesAgri1</a>
**Version:** v1.0.0
