### 2021-06-11 f6ae930
changes in Rd files
