+++ inst/layout/app.R
@@ -68,12 +68,30 @@ library(knitr)
     ##
     tags$br(),
 
-    h5(tags$div(
-      "Package:",
-      tags$br(),
-      tags$b("grapesAgri1, Version 1.0.0"),
-      tags$br()
-    ))
+    h5(
+      tags$div(
+        tags$br(),
+        "Developed by:",
+        tags$br(),
+        tags$b("Dr.Pratheesh P. Gopinath"),
+        tags$br(),
+        tags$b("Assistant Professor,"),
+        tags$br(),
+        tags$b("Agricultural Statistics,"),
+        tags$br(),
+        tags$b("Kerala Agricultural University"),
+        tags$br(),
+        tags$br(),
+        "Contribution:",
+        tags$br(),
+        tags$b("Adarsh V.S."),
+        tags$br(),
+        tags$b("MSc, Agricultural Statistics"),
+        tags$br(),
+        tags$br(),
+        "post your queries at: pratheesh.pg@kau.in"
+        ,style="color:#343aeb")
+    )
 
   ),
   mainPanel(
