+++ inst/crd/app.R
@@ -31,12 +31,30 @@ ui <- fluidPage(
     conditionalPanel("$('#trtmeans').hasClass('recalculating')",
                      tags$div(tags$b('Loading ...please wait while we are calculating in the background.....please dont press submit button again '), style="color:green")),
     tags$br(),
-    h5(tags$div(
-      "Package:",
-      tags$br(),
-      tags$b("grapesAgri1, Version 1.0.0"),
-      tags$br()
-    ))
+    h5(
+      tags$div(
+        tags$br(),
+        "Developed by:",
+        tags$br(),
+        tags$b("Dr.Pratheesh P. Gopinath"),
+        tags$br(),
+        tags$b("Assistant Professor,"),
+        tags$br(),
+        tags$b("Agricultural Statistics,"),
+        tags$br(),
+        tags$b("Kerala Agricultural University"),
+        tags$br(),
+        tags$br(),
+        "Contribution:",
+        tags$br(),
+        tags$b("Adarsh V.S."),
+        tags$br(),
+        tags$b("MSc, Agricultural Statistics"),
+        tags$br(),
+        tags$br(),
+        "post your queries at: pratheesh.pg@kau.in"
+        ,style="color:#343aeb")
+    )
   )
   , mainPanel(
 
