+++ README.md
@@ -11,7 +11,7 @@
 
 
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4923220.svg)](https://doi.org/10.5281/zenodo.4923220)
-
+[![Travis build status](https://travis-ci.com/pratheesh3780/grapesAgri1.svg?branch=master)](https://travis-ci.com/pratheesh3780/grapesAgri1)
 
 
 ## Introduction
