+++ README.md
@@ -19,6 +19,7 @@
 ![GitHub top language](https://img.shields.io/github/languages/top/pratheesh3780/grapesAgri1)
 ![Libraries.io dependency status for GitHub repo](https://img.shields.io/librariesio/github/pratheesh3780/grapesAgri1)
 ![GitHub code size in bytes](https://img.shields.io/github/languages/code-size/pratheesh3780/grapesAgri1)
+[![R-CMD-check](https://github.com/pratheesh3780/grapesAgri1/workflows/R-CMD-check/badge.svg)](https://github.com/pratheesh3780/grapesAgri1/actions)
 
 ---
 
