+++ DESCRIPTION
@@ -20,33 +20,34 @@ Authors@R: c(
            role = c("cph"),
 		       comment = c(url = "http://www.kau.in/")))
 Description: Allows user to have graphical user interface to perform analysis of Agricultural experimental data. On using the functions in this package a Interactive User Interface will pop up. Apps Works by simple upload of files in CSV format.
-License: GPL-3 + file LICENSE
 Encoding: UTF-8
 RoxygenNote: 7.1.1
-Imports: shiny(>= 1.6.0),
-         shinycssloaders(>= 1.0.0),
-         shinyWidgets(>= 0.6.0),
-         rmarkdown(>= 2.7),
-         knitr(>= 1.31),
-         magrittr(>= 2.0.1),
-         dplyr(>= 1.0.4),
-         datasets(>= 4.0.4),
-         gridGraphics(>= 0.5.1),
-         grid(>= 4.0.4),
-         RColorBrewer(>= 1.1.2),
-         Rdpack(>= 2.1.2),
-         devtools(>= 2.4.1),
-         agricolae(>= 1.3.5),
-         pastecs(>= 1.3.21),
-         ggpubr(>= 0.4.0),
-         Hmisc(>= 4.5.0),
-         ggplot2(>= 3.3.3)
+URL: https://github.com/pratheesh3780/grapesAgri1
+    https://CRAN.R-project.org/package=grapesAgri1
+    https://www.kaugrapes.com/
+BugReports: https://github.com/pratheesh3780/grapesAgri1/issues
+Imports: shiny,
+         shinycssloaders,
+         shinyWidgets,
+         rmarkdown,
+         knitr,
+         magrittr,
+         dplyr,
+         datasets,
+         gridGraphics,
+         RColorBrewer,
+         Rdpack,
+         agricolae,
+         pastecs,
+         ggpubr,
+         Hmisc,
+         ggplot2
 Suggests:
-        tinytex(>= 0.32),
-        kableExtra(>= 1.3.4),
-        summarytools(>= 0.9.9),
-        reshape2(>= 1.4.4),
-        desplot(>= 1.8),
-        corrplot(>= 0.84),
-        PairedData(>= 1.1.1)
+        tinytex,
+        kableExtra,
+        summarytools,
+        reshape2,
+        desplot,
+        corrplot,
+        PairedData
 RdMacros: Rdpack
