+++ .travis.yml
@@ -3,3 +3,6 @@
 language: r
 cache: packages
 warnings_are_errors: false
+addons:
+  apt:
+    packages: libgsl0-dev
