+++ man/corrApp.Rd
@@ -17,7 +17,8 @@ like correlogram and scatterplot by uploading CSV file.
 \details{
 This app uses \code{cor.test} to calculate correlation. Correlation matrix
 is calculated using \code{rcorr} function in \code{Hmisc} package. Correlogram
-is obtained using \code{corrplot} function in \code{corrplot} package
+is obtained using \code{corrplot} function in \code{corrplot} package. To download the results in pdf
+one may use \code{tinytex::install_tinytex()} before using the app.
 }
 \examples{
 if(interactive()){
