### 2021-06-30 #3359 @fabian-s
@dlebauer @KristinaRiemer @adithirgis --

would any of you be willing to review this submission for JOSS?

Thank you for considering this!
