### 2021-07-01 pratheesh3780/grapesAgri1#1 @pratheesh3780
Thank you so much for taking your time to give such a detailed review of the package.  All your comments were valuable and will act accordingly.
1. In comment No.2 . grapesAgri1 package is now available in CRAN, should I mention that because It was released in CRAN after the submission of the paper in JOSS.
2. In comment No. 12 . You may select Treatment as FEED and variable of interest (Response) as WEIGHT. Please see if this was the issue.
3. In comment No.11, I was getting a good formatting in Word . Please help me to identify the issue
4. In comment 10. Issue is : To download the results in pdf one may use  `tinytex::install_tinytex()` before using the app. If this may be the issue, I should add it to Readme.
Rest all the comments were clear and will certainly improve the app. I will make the suggested corrections soon.

> Thank you so much
