### 2021-07-01 pratheesh3780/grapesAgri1#1 @adithirgis
> In comment No.2 . grapesAgri1 package is now available in CRAN, should I mention that because It was released in CRAN after the submission of the paper in JOSS.

Has the package been referenced anywhere? Meaning its usage has been in a peer reviewed journal article?

> In comment No. 12 . You may select Treatment as FEED and variable of interest (Response) as WEIGHT. Please see if this was the issue.

Yes, Thanks it works now, could you mention the requirements like such in the ReadMe? Will help users to navigate better.

> In comment No.11, I was getting a good formatting in Word . Please help me to identify the issue

Sorry I get this error in RStudio cloud as well. Could you check using [RStudio cloud](https://rstudio.cloud/)?

> 4. In comment 10. Issue is : To download the results in pdf one may use `tinytex::install_tinytex()` before using the app. If this may be the issue, I should add it to Readme.

Yes please add all the necessary packages in Imports in [DESCRIPTION](https://github.com/adithirgis/grapesAgri1/blob/master/DESCRIPTION).

Also if you are making changes to your paper, please update it by calling whedon.
Cheers!
Adithi
