### 2021-07-01 pratheesh3780/grapesAgri1#1 @adithirgis
Oh sure, please go to [this](https://github.com/openjournals/joss-reviews/issues/3437) page and comment `@whedon commands` or `@whedon generate pdf` to generate pdf version of your paper. I am sorry I made a few changes to the paper.md. Please feel free to accept or reject the formatting change. The strikeout is for removing the words or sentences and the bold is for replacing. I hope you understand.

Yes probably it is the ms office issue, can you please check what would can users do who do not have the latest ms office version?

Thanks!
