### 2021-07-02 pratheesh3780/grapesAgri1#1 @pratheesh3780
1. Can you please add references in paper.md using paper.bib.
 I have already added that, is that ok or any other issue, I didn't understand the comment, sorry

9.When I choose the car.csv and try to use grapesAgri1::descApp() without header it shows me this error -
car.csv already have a header. so if you uncheck header system will consider the first row also as input, which is qualitative and hence the error.

please see this also

Regards!
