### 2021-07-02 pratheesh3780/grapesAgri1#1 @adithirgis
> I have already added that, is that ok or any other issue, I didn't understand the comment, sorry

You will have to add reference inside the paper.md as well, for which ever package you are referring to. Please refer to the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html) of JOSS.

> car.csv already have a header. so if you uncheck header system will consider the first row also as input, which is qualitative and hence the error.

Can you please mention in the ReadMe file or any appropriate place that header is necessary in the file or it creates a confusion. I hope I made sense.

Cheers!
Adithi
