+++ paper.md
@@ -22,7 +22,7 @@ authors:
     affiliation: "3"
     
 affiliations:
- - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural **University**, Kerala, India.
+ - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural University, Kerala, India.
    index: 1
  - name: ICAR-Indian Agricultural Statistics Research Institute, New Delhi, India.
    index: 2
@@ -36,53 +36,33 @@ bibliography: paper.bib
 # Summary
 
 <p>
+Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR for data analysis. Several online platforms were available for data analysis in Agriculture, most of them lack option to generate graphs and plots for publication use. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
+</p>
 
-::: {align="justify"}
-*Can this paragraph's gist be present in the Statement of need where the author can mention the target audience?*
-India has one of the world's largest agricultural education **systems**. According to the Indian Council of Agricultural Research (ICAR), the main regulator of agricultural education in India, there are 63 State Agricultural Universities, 4 deemed universities and 3 Central Agricultural Universities. These institutions enrol on **an** annual basis about 15,000 students at **Undergraduate** and over 7,000 students at Post Graduate and more than 2000 at PhD level. At any point, there are over 75,000 students studying in these institutions. Research activities are performed actively in all these universities.
-:::
+<p>
+R, which is an open source software, provides an excellent platform for data analysis. This powerful computation platform is not much utilised by researchers in the agricultural field. Programming and computational knowledge hinders agricultural researchers from using R for analysis. Also, researchers in Agriculture prefer a graphical user interface.
+</p>
+
+<p>
+Efforts should be made to develop a high quality reliable open-source platform with a simple interactive user interface for data analysis in agriculture. In this light utilising the features of shiny package in R, we have developed a collection of shiny apps for agricultural research called `grapesAgri1` (General R-shiny based Analysis Platform Empowered by Statistics for data analysis in Agriculture-part1). It is hosted on the web (www.kaugrapes.com), downloadable as a standalone application and now we have realised it in to an R package.
+
 
 </p>
 
 <p>
-
-::: {align="justify"}
-Agricultural experiments demand a wide range of statistical tools for analysis, which includes ~~from~~ Exploratory analysis, Design of experiments and Statistical genetics. ~~Scientist and research students were in trouble to find a suitable platform for their data analysis and also to publish the research output in quality journals.~~ **It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals.** Most of the **software** available for data analysis are proprietary or **lack** a simple user interface. Though several online platforms were available for data analysis, they are not much updated up to the current publication standards at international level. Also, these web applications don't provide **opportunities** to generate plots and graphs for publication.
-:::
-
-</p>
-
-<p>
-
-::: {align="justify"}
[…]
