+++ inst/desc/app.R
@@ -12,27 +12,41 @@ library(ggpubr)
 library(summarytools)
 library(ggplot2)
 ############################### ui
-ui <- fluidPage(
+ui = fluidPage(
   setBackgroundColor(
     color = c("#faf1d2", "#ffffff"),
     gradient = "radial",
     direction = c("bottom", "right")
   ),
-  titlePanel(tags$b('Descriptive Statistics & Visualization')),
+  tags$script("$(document).on('shiny:connected', function(event) {
+var myWidth = $(window).width();
+Shiny.onInputChange('shiny_width',myWidth)
+
+});"),
+  tags$script("$(document).on('shiny:connected', function(event) {
+var myHeight = $(window).height();
+Shiny.onInputChange('shiny_height',myHeight)
+
+});"),
+  titlePanel(tags$b("Descriptive Statistics & Visualization")),
   sidebarPanel(
     fileInput("file1", "CSV File (upload in csv format)",
-              accept=c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
+      accept = c("text/csv", "text/comma-separated-values,text/plain", ".csv")
+    ),
     checkboxInput("header", "Header", TRUE),
-    selectInput('req', 'Please select the analysis type',
-                c(SUMMARY= 'summary',
-                  SUMMARY_by_Group = 'sumbygrp',
-                  BOX_PLOT= 'boxplot',
-                  Histogram='histogram',
-                  QQ_Plot='qqplot',
-                  Normality_Test ='nt'
-                )
-                ,'summary')
-    ,uiOutput('var'),
+    selectInput(
+      "req", "Please select the analysis type",
+      c(
+        SUMMARY = "summary",
+        SUMMARY_by_Group = "sumbygrp",
+        BOX_PLOT = "boxplot",
+        Histogram = "histogram",
+        QQ_Plot = "qqplot",
+        Normality_Test = "nt"
+      ),
+      "summary"
+    ),
+    uiOutput("var"),
     tags$br(),
     h5(
       tags$div(
@@ -55,317 +69,399 @@ ui <- fluidPage(
         tags$b("MSc, Agricultural Statistics"),
         tags$br(),
         tags$br(),
-        "post your queries at: pratheesh.pg@kau.in"
-        ,style="color:#343aeb")
+        "post your queries at: pratheesh.pg@kau.in",
+        style = "color:#343aeb"
+      )
     )
-
+  ),
+  mainPanel(
+    htmlOutput("note"),
+    uiOutput("data_set"), # for data set download
+    tableOutput("summaryout"),
+    verbatimTextOutput("nort"),
+    htmlOutput("text3"),
+    tableOutput("bygroup"),
+    uiOutput("var1"),
+    uiOutput("var2"),
+    uiOutput("plot"),
+    tags$br(),
+    uiOutput("image_down")
   )
-  , mainPanel(
-    htmlOutput('note'),
-    uiOutput('data_set'),# for data set download
-    tableOutput('summaryout'),
-   verbatimTextOutput('nort'),
-   htmlOutput('text3'),
-    tableOutput('bygroup'),
-    uiOutput('var1'),
-   uiOutput('var2'),
-   uiOutput('plot'),
-   tags$br(),
-   uiOutput('image_down')
-
-)
 )
 
 #################################### SERVER
 server = function(input, output, session) {
-
-  csvfile <- reactive({
-    csvfile <- input$file1
-    if (is.null(csvfile)){return(NULL)}
-    dt <- read.csv(csvfile$datapath, header=input$header, sep=",")
+  csvfile = reactive({
+    csvfile = input$file1
+    if (is.null(csvfile)) {
+      return(NULL)
+    }
[…]
