+++ README.md
@@ -5,14 +5,35 @@ output:
 ---
 # `grapesAgri1`: Collection of shiny applications for data analysis in Agriculture-Part 1 <img src="man/figures/logo.PNG" align="right" alt="logo" width="173" height = "200" style = "border: none; float: right;">
 ### General R-shiny based Analysis Platform Empowered by Statistics in Agriculture part-1 (grapesAgri1)
+
 ## R-Package for Data Analysis in Agriculture.
 ###### Version : 0.1.0; Copyright (C) 2021-2025: [Kerala Agricultural University](http://www.kau.in/); License: [GPL-3](https://www.r-project.org/Licenses/) 
+=======
+###### Version : 0.1.0; Copyright (C) 2021-2025: [Kerala Agricultural University](https://www.kaugrapes.com); License: [GPL-3](https://www.r-project.org/Licenses/) 
+
 ##### *Gopinath, P. P.<sup>1</sup>, Parsad, R.<sup>2</sup>, Joseph, B.<sup>1</sup>, Adarsh, V.S.<sup>3</sup>*
 
 1.  Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural Univesity.
 2.  ICAR-Indian Agricultural Statistics Research Institute,
     New Delhi.
 3.  Department of Agricultural Statistics, BCKV, West Bengal
+
+---
+
+![CRAN/METACRAN](https://img.shields.io/cran/v/grapesAgri1?style=for-the-badge)
+![GitHub](https://img.shields.io/github/license/pratheesh3780/grapesAgri1)
+[![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version-last-release/grapesAgri1)](https://cran.r-project.org/package=grapesAgri1)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4923220.svg)](https://doi.org/10.5281/zenodo.4923220)
+![GitHub release (latest by date)](https://img.shields.io/github/v/release/pratheesh3780/grapesAgri1)
+![GitHub R package version](https://img.shields.io/github/r-package/v/pratheesh3780/grapesAgri1)
+![GitHub language count](https://img.shields.io/github/languages/count/pratheesh3780/grapesAgri1)
+![Libraries.io dependency status for GitHub repo](https://img.shields.io/librariesio/github/pratheesh3780/grapesAgri1)
+![GitHub code size in bytes](https://img.shields.io/github/languages/code-size/pratheesh3780/grapesAgri1)
+[![R-CMD-check](https://github.com/pratheesh3780/grapesAgri1/workflows/R-CMD-check/badge.svg)](https://github.com/pratheesh3780/grapesAgri1/actions)
+[![](https://cranlogs.r-pkg.org/badges/grapesAgri1)](https://cran.r-project.org/package=grapesAgri1)
+[![Build Status](https://www.travis-ci.com/pratheesh3780/grapesAgri1.svg?branch=master)](https://www.travis-ci.com/pratheesh3780/grapesAgri1)
+
+---
 
 ## Introduction
[…]
