+++ paper.md
@@ -33,21 +33,16 @@ bibliography: paper.bib
 # Summary
 
 <p>
-Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR for data analysis. Several online platforms were available for data analysis in Agriculture, most of them lack option to generate graphs and plots for publication use. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
+Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR (Indian Council of Agricultural Research) for data analysis, though it is a highly advanced statistical analysis platform, its complexity hold back students and researchers from using it. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
 </p>
-
 <p>
 R, which is an open source software, provides an excellent platform for data analysis. This powerful computation platform is not much utilised by researchers in the agricultural field. Programming and computational knowledge hinders agricultural researchers from using R for analysis. Also, researchers in Agriculture prefer a graphical user interface.
 </p>
-
 <p>
-Efforts should be made to develop a high quality reliable open-source platform with a simple interactive user interface for data analysis in agriculture. In this light utilising the features of shiny package in R, we have developed a collection of shiny apps for agricultural research called `grapesAgri1` (General R-shiny based Analysis Platform Empowered by Statistics for data analysis in Agriculture-part1). It is hosted on the web (www.kaugrapes.com), downloadable as a standalone application and now we have realised it in to an R package.
-
-
+Efforts should be made to develop a high quality, reliable open-source platform with a simple interactive user interface for data analysis in agriculture. In this light utilising the features of shiny package in R, we have developed a collection of shiny apps for agricultural research called `grapesAgri1` (General R-shiny based Analysis Platform Empowered by Statistics for data analysis in Agriculture-part1). It is hosted on the web (www.kaugrapes.com), downloadable as a standalone application and now we have realised it in to an R package.
 </p>
-
 <p>
-On using the functions in `grapesAgri1` package a Graphical User Interface will pop up. Apps are self-explanatory and works by simple upload of files in CSV format. Results can be downloaded as docx, PDF or HTML format. Plots and Graphs can be generated, which are also downloadable as .png file.
+On using the functions in `grapesAgri1` package a Graphical User Interface will pop up. Apps are self-explanatory and works by simple upload of files in CSV format. Results can be downloaded in HTML format. Plots and Graphs can be generated, which are also downloadable as .png file.
 </p>
 
 # Statement of need
[…]
