+++ inst/desc/tests/shinytest/mytest.R
@@ -1,26 +1,0 @@
-app <- ShinyDriver$new(loadTimeout = 100000)
-app$snapshotInit("mytest")
-
-app$snapshotDownload("downloadData")
-app$uploadFile(file1 = "iris.csv") # <-- This should be the path to the file, relative to the app's tests/shinytest directory
-app$setInputs(var = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(submit4 = "click")
-app$snapshot()
-app$setInputs(req = "sumbygrp")
-app$setInputs(var = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(group = "Species",allowInputNoBinding_ = TRUE)
-app$setInputs(submit5 = "click")
-app$snapshot()
-app$setInputs(req = "boxplot")
-app$setInputs(variable = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(submit1 = "click")
-app$snapshot()
-app$setInputs(req = "histogram")
-app$setInputs(submit2 = "click")
-app$snapshot()
-app$setInputs(req = "qqplot")
-app$setInputs(submit3 = "click")
-app$snapshot()
-app$setInputs(req = "nt")
-app$setInputs(submit = "click")
-app$snapshot()
