+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(grapesAgri1)
+
+test_check("grapesAgri1")
