+++ tests/testthat/testdesc.R
@@ -1,0 +1,8 @@
+context("app-file")
+library(shinytest)
+context("app-file")
+test_that("descApp_testing", {
+  skip_on_cran()
+  appDir <- system.file("desc", package = "grapesAgri1")
+  expect_pass(testApp(appDir, compareImages = FALSE))
+})
