+++ tests/testthat/testdesc.R
@@ -1,6 +1,5 @@
 context("app-file")
 library(shinytest)
-context("app-file")
 test_that("descApp_testing", {
   skip_on_cran()
   appDir <- system.file("desc", package = "grapesAgri1")
