+++ DESCRIPTION
@@ -27,28 +27,32 @@ URL: https://github.com/pratheesh3780/grapesAgri1
     https://CRAN.R-project.org/package=grapesAgri1
     https://www.kaugrapes.com/
 BugReports: https://github.com/pratheesh3780/grapesAgri1/issues
-Imports: shiny,
-         shinycssloaders,
-         shinyWidgets,
-         rmarkdown,
-         knitr,
-         magrittr,
-         dplyr,
-         datasets,
-         gridGraphics,
-         RColorBrewer,
-         Rdpack,
-         agricolae,
-         pastecs,
-         ggpubr,
-         Hmisc,
-         ggplot2
+Imports: 
+        shiny(>= 1.6.0),
+        shinycssloaders(>= 1.0.0),
+        shinyWidgets(>= 0.6.0),
+        rmarkdown(>= 2.7),
+        knitr(>= 1.31),
+        kableExtra(>= 1.3.4),
+        magrittr(>= 2.0.1),
+        summarytools(>= 0.9.9),
+        dplyr(>= 1.0.4),
+        pastecs(>= 1.3.21),
+        ggpubr(>= 0.4.0),
+        Hmisc(>= 4.5.0),
+        datasets(>= 4.0.4),
+        corrplot(>= 0.84),
+        ggplot2(>= 3.3.3),
+        reshape2(>= 1.4.4),
+        gridGraphics(>= 0.5.1),
+        grid(>= 4.0.4),
+        RColorBrewer(>= 1.1.2),
+        desplot(>= 1.8),
+        agricolae(>= 1.3.5),
+        PairedData(>= 1.1.1),
+        Rdpack(>= 2.1.2)
 Suggests:
-        tinytex,
-        kableExtra,
-        summarytools,
-        reshape2,
-        desplot,
-        corrplot,
-        PairedData
-RdMacros: Rdpack
+        shinytest,
+        testthat
+RdMacros:
+        Rdpack
