+++ inst/comp_mean/app.R
@@ -10,7 +10,7 @@ library(ggpubr)
 library(PairedData)
 library(reshape2)
 ############################### Ui
-ui <- fluidPage(
+ui = fluidPage(
   setBackgroundColor(
     color = c("#ffb9b3", "#ffffff"),
     gradient = "radial",
@@ -195,13 +195,13 @@ ui <- fluidPage(
   )
 )
 ########################### Server
-server <- function(input, output, session) {
-  csvfile <- reactive({
-    csvfile <- input$file1
+server = function(input, output, session) {
+  csvfile = reactive({
+    csvfile = input$file1
     if (is.null(csvfile)) {
       return(NULL)
     }
-    dt <-
+    dt =
       read.csv(csvfile$datapath,
         header = input$header,
         sep = ","
@@ -209,7 +209,7 @@ server <- function(input, output, session) {
     dt
   })
 
-  output$var <- renderUI({
+  output$var = renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -378,7 +378,7 @@ server <- function(input, output, session) {
     }
   })
   # two sample ttest
-  output$ttest <- renderTable(
+  output$ttest = renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -391,9 +391,9 @@ server <- function(input, output, session) {
       }
       if (input$req1 == "ttest") {
         if (input$submit1 > 0) {
-          x <- as.vector(csvfile()[, input$dvar])
-          y <- as.vector(csvfile()[, input$ivar])
-          t <- t.test(
+          x = as.vector(csvfile()[, input$dvar])
+          y = as.vector(csvfile()[, input$ivar])
+          t = t.test(
             x,
             y,
             alternative = input$alt,
@@ -401,12 +401,12 @@ server <- function(input, output, session) {
             paired = FALSE,
             na.omit = TRUE
           )
-          t_value <- round(t$statistic, 3)
-          df <- t$parameter
-          Pvalue <- round(t$p.value, 3)
-          alt.Hypothesis <- t$alternative
-          result <- cbind(t_value, df, Pvalue, alt.Hypothesis)
-          result <- as.data.frame(result)
+          t_value = round(t$statistic, 3)
+          df = t$parameter
+          Pvalue = round(t$p.value, 3)
+          alt.Hypothesis = t$alternative
+          result = cbind(t_value, df, Pvalue, alt.Hypothesis)
+          result = as.data.frame(result)
           result
         }
       }
@@ -419,7 +419,7 @@ server <- function(input, output, session) {
   )
 
   # note1
-  output$note1 <- renderUI({
+  output$note1 = renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -431,9 +431,9 @@ server <- function(input, output, session) {
     }
     if (input$req1 == "ttest") {
       if (input$submit1 > 0) {
-        x <- as.vector(csvfile()[, input$dvar])
-        y <- as.vector(csvfile()[, input$ivar])
-        t <- t.test(
+        x = as.vector(csvfile()[, input$dvar])
+        y = as.vector(csvfile()[, input$ivar])
+        t = t.test(
           x,
           y,
           alternative = input$alt,
@@ -462,7 +462,7 @@ server <- function(input, output, session) {
       }
     }
   })
-  output$summary1 <- renderTable(
+  output$summary1 = renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -475,16 +475,16 @@ server <- function(input, output, session) {
       }
       if (input$req1 == "ttest") {
         if (input$submit1 > 0) {
-          grp1 <- subset(csvfile(), select = input$dvar)
-          grp2 <- subset(csvfile(), select = input$ivar)
[…]
