+++ inst/comp_mean/tests/shinytest/mytest.R
@@ -1,0 +1,18 @@
+app <- ShinyDriver$new("../../")
+app$snapshotInit("mytest")
+
+app$snapshot()
+app$uploadFile(file1 = "iris.csv") # <-- This should be the path to the file, relative to the app's tests/shinytest directory
+app$setInputs(submit1 = "click")
+app$setInputs(req1 = "pttest")
+app$setInputs(submit5 = "click")
+app$setInputs(req1 = "ottest")
+app$setInputs(submit3 = "click")
+app$setInputs(req1 = "wttest")
+app$setInputs(submit2 = "click")
+app$setInputs(req1 = "ftest")
+app$setInputs(submit4 = "click")
+app$setInputs(req1 = "boxplot")
+app$setInputs(submit6 = "click")
+app$setInputs(req1 = "pplot")
+app$setInputs(submit7 = "click")
