+++ inst/comp_mean/app.R
@@ -1,6 +1,5 @@
 library(shiny)
 library(shinyWidgets)
-library(shinycssloaders)
 library(datasets)
 library(rmarkdown)
 library(dplyr)
@@ -10,7 +9,7 @@ library(ggpubr)
 library(PairedData)
 library(reshape2)
 ############################### Ui
-ui = fluidPage(
+ui <- fluidPage(
   setBackgroundColor(
     color = c("#ffb9b3", "#ffffff"),
     gradient = "radial",
@@ -195,13 +194,13 @@ ui = fluidPage(
   )
 )
 ########################### Server
-server = function(input, output, session) {
-  csvfile = reactive({
-    csvfile = input$file1
+server <- function(input, output, session) {
+  csvfile <- reactive({
+    csvfile <- input$file1
     if (is.null(csvfile)) {
       return(NULL)
     }
-    dt =
+    dt <-
       read.csv(csvfile$datapath,
         header = input$header,
         sep = ","
@@ -209,7 +208,7 @@ server = function(input, output, session) {
     dt
   })
 
-  output$var = renderUI({
+  output$var <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -309,7 +308,7 @@ server = function(input, output, session) {
     else if (input$req1 == "pttest") {
       list(
         radioButtons("dvar", "Please pick 'Group A' ", choices = names(csvfile())),
-        radioButtons("ivar", "Please pick 'Group B'", choices = names(csvfile())),
+        radioButtons("ivar", "Please pick 'Group B'", choices = names(csvfile()), selected = NULL),
         selectInput(
           "alt",
           "Alternative Hypotesis",
@@ -378,7 +377,7 @@ server = function(input, output, session) {
     }
   })
   # two sample ttest
-  output$ttest = renderTable(
+  output$ttest <- renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -391,9 +390,9 @@ server = function(input, output, session) {
       }
       if (input$req1 == "ttest") {
         if (input$submit1 > 0) {
-          x = as.vector(csvfile()[, input$dvar])
-          y = as.vector(csvfile()[, input$ivar])
-          t = t.test(
+          x <- as.vector(csvfile()[, input$dvar])
+          y <- as.vector(csvfile()[, input$ivar])
+          t <- t.test(
             x,
             y,
             alternative = input$alt,
@@ -401,12 +400,12 @@ server = function(input, output, session) {
             paired = FALSE,
             na.omit = TRUE
           )
-          t_value = round(t$statistic, 3)
-          df = t$parameter
-          Pvalue = round(t$p.value, 3)
-          alt.Hypothesis = t$alternative
-          result = cbind(t_value, df, Pvalue, alt.Hypothesis)
-          result = as.data.frame(result)
+          t_value <- round(t$statistic, 3)
+          df <- t$parameter
+          Pvalue <- round(t$p.value, 3)
+          alt.Hypothesis <- t$alternative
+          result <- cbind(t_value, df, Pvalue, alt.Hypothesis)
+          result <- as.data.frame(result)
           result
         }
       }
@@ -419,7 +418,7 @@ server = function(input, output, session) {
   )
 
   # note1
-  output$note1 = renderUI({
+  output$note1 <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -431,9 +430,9 @@ server = function(input, output, session) {
     }
     if (input$req1 == "ttest") {
       if (input$submit1 > 0) {
-        x = as.vector(csvfile()[, input$dvar])
-        y = as.vector(csvfile()[, input$ivar])
-        t = t.test(
+        x <- as.vector(csvfile()[, input$dvar])
[…]
