+++ tests/testthat/testttest.R
@@ -1,0 +1,7 @@
+context("app-file")
+library(shinytest)
+test_that("descApp_testing", {
+  skip_on_cran()
+  appDir <- system.file("comp_mean", package = "grapesAgri1")
+  expect_pass(testApp(appDir, compareImages = FALSE))
+})
