+++ R/ttest.R
@@ -2,7 +2,7 @@
 #' @return Nothing
 #' @description
 #' ttApp() function opens up an interactive shiny app which will allow
-#' user to easily perform one sample t-test, unpaired two sample t-test
+#' user to easily perform one sample t-test, unpaired two sample t-test,
 #' unpaired two sample Welch t-test, paired t-test, test for homogeneity of variance (F-test),
 #' and obtain plots like boxplot and paired plot by uploading CSV file.
 #' @details
