+++ R/crd.R
@@ -1,21 +1,20 @@
-#'@title Completely Randomized Design
-#'@return Nothing
-#'@description
-#'crdApp() function opens up an interactive shiny app which will allow
-#'the user to perform analysis of completely randomized design with
-#'equal or unequal replications. Multiple comparison tests like LSD,
-#'DMRT and Tukey can be performed. Box-pot and Bar-chart with confidence interval
-#'can be plotted. All these can be achieved by uploading CSV file.
+#' @title Completely Randomized Design
+#' @return Nothing
+#' @description
+#' crdApp() function opens up an interactive shiny app which will allow
+#' the user to perform analysis of completely randomized design with
+#' equal or unequal replications. Multiple comparison tests like LSD,
+#' DMRT and Tukey can be performed. Box-pot and Bar-chart with confidence interval
+#' can be plotted. All these can be achieved by uploading CSV file.
 #'
 #' @details
-#'This app uses \code{anova} function of \code{stats} package to
-#'obtain one-way ANOVA.\code{LSD.test},\code{duncan.test} and
-#'\code{HSD.test} functions of \code{agricolae} package is used for
-#'multiple comparison test like LSD,DMRT and Tukey respectively.
-#'\code{ggboxplot} function of \code{ggpubr} package is used for
-#'boxplot.'\code{ggplot} function of \code{ggplot2} is used for
-#'barchart with confidence interval. To download the results in pdf
-#'one may use \code{tinytex::install_tinytex()} before using the app.
+#' This app uses \code{anova} function of \code{stats} package to
+#' obtain one-way ANOVA.\code{LSD.test},\code{duncan.test} and
+#' \code{HSD.test} functions of \code{agricolae} package is used for
+#' multiple comparison test like LSD,DMRT and Tukey respectively.
+#' \code{ggboxplot} function of \code{ggpubr} package is used for
+#' boxplot.'\code{ggplot} function of \code{ggplot2} is used for
+#' barchart with confidence interval.
 #'
 #' @keywords
 #' One-way ANOVA
@@ -33,43 +32,40 @@
 #' @export
 #'
 #' @examples
-#'if(interactive()){
-#'crdApp()
-#'}
+#' if (interactive()) {
+#'   crdApp()
+#' }
+#' @references
 #'
 #'
-#'@references
+#' \insertRef{R_2021}{grapesAgri1}
 #'
+#' \insertRef{shiny_2021}{grapesAgri1}
 #'
-#'\insertRef{R_2021}{grapesAgri1}
+#' \insertRef{sw_2021}{grapesAgri1}
 #'
-#'\insertRef{shiny_2021}{grapesAgri1}
+#' \insertRef{dplyr_2021}{grapesAgri1}
 #'
-#'\insertRef{sw_2021}{grapesAgri1}
+#' \insertRef{ggpubr_2020}{grapesAgri1}
 #'
-#'\insertRef{dplyr_2021}{grapesAgri1}
+#' \insertRef{ggplot_2016}{grapesAgri1}
 #'
-#'\insertRef{ggpubr_2020}{grapesAgri1}
+#' \insertRef{gupta1985statistical}{grapesAgri1}
 #'
-#'\insertRef{ggplot_2016}{grapesAgri1}
+#' \insertRef{tukey1977exploratory}{grapesAgri1}
 #'
-#'\insertRef{gupta1985statistical}{grapesAgri1}
+#' \insertRef{hmisc_2021}{grapesAgri1}
 #'
-#'\insertRef{tukey1977exploratory}{grapesAgri1}
[…]
