+++ inst/crd/app.R
@@ -21,15 +21,16 @@ ui <- fluidPage(
   ),
 
   ########
-  titlePanel(tags$div(tags$b('Completely Randomized Design (One-way ANOVA)',style="color:#000000"))),
-
+  titlePanel(tags$div(tags$b("Completely Randomized Design (One-way ANOVA)", style = "color:#000000"))),
   sidebarPanel(
-    fileInput("file1", "CSV File (upload in csv format)", accept=c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
+    fileInput("file1", "CSV File (upload in csv format)", accept = c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
     checkboxInput("header", "Header", TRUE),
-    uiOutput('var'),
+    uiOutput("var"),
     tags$br(),
-    conditionalPanel("$('#trtmeans').hasClass('recalculating')",
-                     tags$div(tags$b('Loading ...please wait while we are calculating in the background.....please dont press submit button again '), style="color:green")),
+    conditionalPanel(
+      "$('#trtmeans').hasClass('recalculating')",
+      tags$div(tags$b("Loading ...please wait while we are calculating in the background.....please dont press submit button again "), style = "color:green")
+    ),
     tags$br(),
     h5(
       tags$div(
@@ -52,490 +53,653 @@ ui <- fluidPage(
         tags$b("MSc, Agricultural Statistics"),
         tags$br(),
         tags$br(),
-        "post your queries at: pratheesh.pg@kau.in"
-        ,style="color:#343aeb")
+        "post your queries at: pratheesh.pg@kau.in",
+        style = "color:#343aeb"
+      )
     )
-  )
-  , mainPanel(
-
-    tabsetPanel(type = "tab",
-                tabPanel("Analysis.Results",
-                         conditionalPanel("$('#trtmeans').hasClass('recalculating')",
-                                          tags$div(tags$b('Loading ...please wait while we are calculating in the background.....please dont press submit button again '), style="color:green")),
-                         tableOutput('trtmeans'),
-                         tags$style(  type="text/css", "#trtmeans th,td {border: medium solid #000000;text-align:center}"),
-                         tags$style(  type="text/css", "#trtmeans td {border: medium solid #000000;text-align:center}"),
-
-                         tableOutput('aovSummary'),
-                         tags$style(  type="text/css", "#aovSummary th,td {border: medium solid #000000;text-align:center}"),
-                         tags$style(  type="text/css", "#aovSummary td {border: medium solid #000000;text-align:center}"),
-
-                         tableOutput('SEM'),
-                         tags$style(  type="text/css", "#SEM th,td {border: medium solid #000000;text-align:center}"),
-                         tags$style(  type="text/css", "#SEM td {border: medium solid #000000;text-align:center}"),
-
-                         h3(""),
-                         htmlOutput('text'),
-                         h3(""),
-
-                         uiOutput('inference'),
-
-                         tableOutput('multi'),
-                         tags$style(  type="text/css", "#multi th,td {border: medium solid #000000;text-align:center}"),
-                         tags$style(  type="text/css", "#multi td {border: medium solid #000000;text-align:center}"),
-
-                         tableOutput('unequal'),
-                         tags$style(  type="text/css", "#unequal th,td {border: medium solid #000000;text-align:center}"),
-                         tags$style(  type="text/css", "#unequal td {border: medium solid #000000;text-align:center}"),
-
-                         tableOutput('group'),
[…]
