+++ inst/Corr/app.R
@@ -1,4 +1,3 @@
-library(shinycssloaders)
 library(rmarkdown)
 library(shinyWidgets)
 library(dplyr)
@@ -10,6 +9,8 @@ library(corrplot)
 library(Hmisc)
 library(reshape2)
 library(ggplot2)
+library(datasets)
+library(grid)
 
 
 
