+++ DESCRIPTION
@@ -29,7 +29,6 @@ URL: https://github.com/pratheesh3780/grapesAgri1
 BugReports: https://github.com/pratheesh3780/grapesAgri1/issues
 Imports: 
         shiny(>= 1.6.0),
-        shinycssloaders(>= 1.0.0),
         shinyWidgets(>= 0.6.0),
         rmarkdown(>= 2.7),
         knitr(>= 1.31),
@@ -40,12 +39,10 @@ Imports:
         pastecs(>= 1.3.21),
         ggpubr(>= 0.4.0),
         Hmisc(>= 4.5.0),
-        datasets(>= 4.0.4),
         corrplot(>= 0.84),
         ggplot2(>= 3.3.3),
         reshape2(>= 1.4.4),
         gridGraphics(>= 0.5.1),
-        grid(>= 4.0.4),
         RColorBrewer(>= 1.1.2),
         desplot(>= 1.8),
         agricolae(>= 1.3.5),
@@ -53,6 +50,8 @@ Imports:
         Rdpack(>= 2.1.2)
 Suggests:
         shinytest,
-        testthat
+        testthat,
+        datasets,
+        grid
 RdMacros:
         Rdpack
