+++ inst/desc/app.R
@@ -13,7 +13,7 @@ library(ggplot2)
 library(datasets)
 library(grid)
 ############################### ui
-ui = fluidPage(
+ui <- fluidPage(
   setBackgroundColor(
     color = c("#faf1d2", "#ffffff"),
     gradient = "radial",
@@ -91,17 +91,17 @@ Shiny.onInputChange('shiny_height',myHeight)
 )
 
 #################################### SERVER
-server = function(input, output, session) {
-  csvfile = reactive({
-    csvfile = input$file1
+server <- function(input, output, session) {
+  csvfile <- reactive({
+    csvfile <- input$file1
     if (is.null(csvfile)) {
       return(NULL)
     }
-    dt = read.csv(csvfile$datapath, header = input$header, sep = ",")
+    dt <- read.csv(csvfile$datapath, header = input$header, sep = ",", check.names = FALSE)
     dt
   })
 
-  output$var = renderUI({
+  output$var <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -241,7 +241,7 @@ server = function(input, output, session) {
   })
 
 
-  output$var1 = renderUI({
+  output$var1 <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -263,7 +263,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$var2 = renderUI({
+  output$var2 <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -286,7 +286,7 @@ server = function(input, output, session) {
   })
 
   ############################# Summary
-  output$summaryout = function() {
+  output$summaryout <- function() {
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -298,8 +298,8 @@ server = function(input, output, session) {
     }
     if (input$req == "summary") {
       if (input$submit4 > 0) {
-        y = subset(csvfile(), select = input$var)
-        final =
+        y <- subset(csvfile(), select = input$var)
+        final <-
           summarytools::descr(y) %>%
           summarytools::tb(order = 3) %>%
           knitr::kable(digits = 2, caption = "Summary Statistics") %>%
@@ -310,7 +310,7 @@ server = function(input, output, session) {
     }
   }
   ####################################### Shapiro Wilk's Test
-  output$nort = renderPrint({
+  output$nort <- renderPrint({
     if (is.null(input$file1$datapath)) {
       return(invisible())
     }
@@ -322,16 +322,16 @@ server = function(input, output, session) {
     }
     if (input$req == "nt") {
       if (input$submit > 0) {
-        y = subset(csvfile(), select = input$var)
-        data = as.data.frame(y)
-        colnames(data) = "variable_under_study"
-        test = stats::shapiro.test(data$variable_under_study)
+        y <- subset(csvfile(), select = input$var)
+        data <- as.data.frame(y)
+        colnames(data) <- "variable_under_study"
+        test <- stats::shapiro.test(data$variable_under_study)
         test
       }
     }
   })
   ############################# Summary by group
-  output$bygroup = function() {
+  output$bygroup <- function() {
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -343,9 +343,9 @@ server = function(input, output, session) {
     }
     if (input$req == "sumbygrp") {
       if (input$submit5 > 0) {
-        y1 = subset(csvfile(), select = input$var)
-        y2 = subset(csvfile(), select = input$group)
-        final =
+        y1 <- subset(csvfile(), select = input$var)
+        y2 <- subset(csvfile(), select = input$group)
+        final <-
           summarytools::stby(y1, y2, descr) %>%
           summarytools::tb(order = 1) %>%
           knitr::kable(digits = 2, caption = "Summary Statistics by Group") %>%
[…]
