+++ inst/Corr/app.R
@@ -4,7 +4,6 @@ library(dplyr)
 library(gridGraphics)
 library(RColorBrewer)
 library(knitr)
-##############
 library(corrplot)
 library(Hmisc)
 library(reshape2)
@@ -17,29 +16,37 @@ library(grid)
 ############################### ui
 
 ui <- fluidPage(
-
   setBackgroundColor(
     color = c("#bbedfc", "#ffffff"),
     gradient = "radial",
     direction = c("bottom", "right")
   ),
+  tags$script("$(document).on('shiny:connected', function(event) {
+var myWidth = $(window).width();
+Shiny.onInputChange('shiny_width',myWidth)
 
-  titlePanel(tags$div(tags$b('Correlation Analysis',style="color:#fc03f4"))),
+});"),
+  tags$script("$(document).on('shiny:connected', function(event) {
+var myHeight = $(window).height();
+Shiny.onInputChange('shiny_height',myHeight)
 
+});"),
+  titlePanel(tags$div(tags$b("Correlation Analysis", style = "color:#fc03f4"))),
   sidebarPanel(
-    fileInput("file1", "CSV File (upload in csv format)", accept=c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
+    fileInput("file1", "CSV File (upload in csv format)", accept = c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
     checkboxInput("header", "Header", TRUE),
-    selectInput('req1', 'Please select the analysis type',
-                c(Simple_correlation = 'correlation',
-                  Correlation_Matrix = 'corrmat',
-                  Scatter_Plot= 'scatplot',
-                  Correlogram ='corrplot'
-                )
-                ,'correlation'),
-    uiOutput('var'),
-
+    selectInput(
+      "req1", "Please select the analysis type",
+      c(
+        Simple_correlation = "correlation",
+        Correlation_Matrix = "corrmat",
+        Scatter_Plot = "scatplot",
+        Correlogram = "corrplot"
+      ),
+      "correlation"
+    ),
+    uiOutput("var"),
     tags$br(),
-
     h5(
       tags$div(
         tags$br(),
@@ -61,381 +68,540 @@ ui <- fluidPage(
         tags$b("MSc, Agricultural Statistics"),
         tags$br(),
         tags$br(),
-        "post your queries at: pratheesh.pg@kau.in"
-        ,style="color:#343aeb")
+        "post your queries at: pratheesh.pg@kau.in",
+        style = "color:#343aeb"
+      )
     )
+  ),
+  mainPanel(
+    htmlOutput("note1"),
+    uiOutput("data_set"), # for data set download
+
+    tableOutput("correlation"),
+    tags$style(type = "text/css", "#correlation th,td {border: medium solid #000000;text-align:center}"),
+    tags$style(type = "text/css", "#correlation td {border: medium solid #000000;text-align:center}"),
+    tableOutput("ci"),
+    tags$style(type = "text/css", "#ci th,td {border: medium solid #000000;text-align:center}"),
+    tags$style(type = "text/css", "#ci td {border: medium solid #000000;text-align:center}"),
+    tableOutput("corrmat"),
+    tags$style(type = "text/css", "#corrmat th,td {border: medium solid #000000;text-align:center}"),
+    tags$style(type = "text/css", "#corrmat td {border: medium solid #000000;text-align:center}"),
+    htmlOutput("note"),
+    tags$br(),
+    tableOutput("sigmat"),
+    tags$style(type = "text/css", "#sigmat th,td {border: medium solid #000000;text-align:center}"),
[…]
