+++ inst/Corr/app.R
@@ -15,7 +15,7 @@ library(grid)
 
 ############################### ui
 
-ui <- fluidPage(
+ui = fluidPage(
   setBackgroundColor(
     color = c("#bbedfc", "#ffffff"),
     gradient = "radial",
@@ -99,17 +99,17 @@ Shiny.onInputChange('shiny_height',myHeight)
   )
 )
 ############################ SERVER
-server <- function(input, output, session) {
-  csvfile <- reactive({
-    csvfile <- input$file1
+server = function(input, output, session) {
+  csvfile = reactive({
+    csvfile = input$file1
     if (is.null(csvfile)) {
       return(NULL)
     }
-    dt <- read.csv(csvfile$datapath, header = input$header, sep = ",", check.names = FALSE)
+    dt = read.csv(csvfile$datapath, header = input$header, sep = ",", check.names = FALSE)
     dt
   })
 
-  output$var <- renderUI({
+  output$var = renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -283,7 +283,7 @@ server <- function(input, output, session) {
 
   ###################
   # correlaton
-  output$correlation <- renderTable(
+  output$correlation = renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -299,21 +299,21 @@ server <- function(input, output, session) {
           validate(
             need(input$dvar != input$ivar, "Both input variables selected (x and y) are same, correlation will be one.")
           )
-          a <- as.vector(csvfile()[, input$dvar])
-          y <- as.vector(csvfile()[, input$ivar])
-          x <- cor.test(a, y,
+          a = as.vector(csvfile()[, input$dvar])
+          y = as.vector(csvfile()[, input$ivar])
+          x = cor.test(a, y,
             method = input$req, conf.level = as.numeric(input$ci),
             alternative = input$alt, exact = FALSE
           )
-          t_value <- round(x$statistic, 3)
-          correlation <- round(x$estimate, 3)
-          df <- x$parameter
-          pvalue <- round(x$p.value, 3)
-          alt.Hypothesis <- x$alternative
-          result <- cbind(correlation, t_value, df, pvalue, alt.Hypothesis)
-          nam <- x$method
-          rownames(result) <- nam
-          result <- as.data.frame(result)
+          t_value = round(x$statistic, 3)
+          correlation = round(x$estimate, 3)
+          df = x$parameter
+          pvalue = round(x$p.value, 3)
+          alt.Hypothesis = x$alternative
+          result = cbind(correlation, t_value, df, pvalue, alt.Hypothesis)
+          nam = x$method
+          rownames(result) = nam
+          result = as.data.frame(result)
           result
         }
       }
@@ -325,7 +325,7 @@ server <- function(input, output, session) {
     caption.placement = getOption("xtable.caption.placement", "top")
   )
 
-  output$ci <- renderTable(
+  output$ci = renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -343,14 +343,14 @@ server <- function(input, output, session) {
               need(input$dvar != input$ivar, "Are you confused?.
             Correlation measures linear relationship between two variables say x and y. To know more on correlation see Read me of package grapesAgri1. You can select different variables as x and y from the sidebar panel and see how it works!")
             )
-            a <- as.vector(csvfile()[, input$dvar])
-            y <- as.vector(csvfile()[, input$ivar])
-            x <- cor.test(a, y,
+            a = as.vector(csvfile()[, input$dvar])
+            y = as.vector(csvfile()[, input$ivar])
+            x = cor.test(a, y,
               method = "pearson", conf.level = as.numeric(input$ci),
[…]
