+++ inst/Corr/report.Rmd
@@ -9,86 +9,101 @@ knitr::opts_chunk$set(echo = FALSE)
 library(knitr)
 library(corrplot)
 library(Hmisc)
+library(magrittr)
+library(kableExtra)
 ```
 ## RESULT
 CORRELATION ANALYSIS 
 
 ```{r, echo = FALSE}
-csvfile <- reactive({
-    csvfile <- input$file1
-    if (is.null(csvfile)){return(NULL)}
-    dt <- read.csv(csvfile$datapath, header=input$header, sep=",",check.names = FALSE)
-    dt
-  })
+csvfile = reactive({
+  csvfile = input$file1
+  if (is.null(csvfile)) {
+    return(NULL)
+  }
+  dt = read.csv(csvfile$datapath, header = input$header, sep = ",", check.names = FALSE)
+  dt
+})
 
-#output
-if(input$req1 == 'correlation'){
-       if(input$submit > 0){
-        a<-as.vector(csvfile()[,input$dvar])
-        y<-as.vector(csvfile()[,input$ivar])
-        x<-cor.test(a, y,method=input$req,conf.level = as.numeric(input$ci)
-                    ,alternative = input$alt,exact = FALSE)
-        t_value<-round(x$statistic,3)
-        correlation<-round(x$estimate,3)
-        df<-x$parameter
-        pvalue<-round(x$p.value,3)
-        alt.Hypothesis<-x$alternative
-        result<-cbind(correlation,t_value,df,pvalue,alt.Hypothesis)
-        nam<-x$method
-        rownames(result)<-nam
-        result<-as.data.frame(result)
-          kable(result,caption = "Correlation Analysis",row.names = FALSE)
-        }
+# output
+if (input$req1 == "correlation") {
+  if (input$submit > 0) {
+    a = as.vector(csvfile()[, input$dvar])
+    y = as.vector(csvfile()[, input$ivar])
+    x = cor.test(a, y,
+      method = input$req, conf.level = as.numeric(input$ci),
+      alternative = input$alt, exact = FALSE
+    )
+    t_value = round(x$statistic, 3)
+    correlation = round(x$estimate, 3)
+    df = x$parameter
+    pvalue = round(x$p.value, 3)
+    alt.Hypothesis = x$alternative
+    result = cbind(correlation, t_value, df, pvalue, alt.Hypothesis)
+    nam = x$method
+    rownames(result) = nam
+    result = as.data.frame(result)
+    kable(result, caption = "Correlation Analysis", row.names = FALSE) %>%
+      kable_styling() %>%
+      kable_paper("hover", full_width = F)
+  }
 }
 
-if(input$req1 == 'correlation'){
-      if(input$submit > 0){
-         if(input$req =='pearson'){
-        a<-as.vector(csvfile()[,input$dvar])
-        y<-as.vector(csvfile()[,input$ivar])
-        x<-cor.test(a, y,method=input$req,conf.level = as.numeric(input$ci)
-                    ,alternative = input$alt,exact = FALSE)
-        ci<-x$conf.int
-        ci_nw = melt(ci, value.name = "Lower Limit and Upper limit")
-        kable(round(ci_nw,3),caption = "Confidence Interval",row.names = FALSE)
-         }
-      }
-    } 
+if (input$req1 == "correlation") {
+  if (input$submit > 0) {
+    if (input$req == "pearson") {
+      a = as.vector(csvfile()[, input$dvar])
+      y = as.vector(csvfile()[, input$ivar])
+      x = cor.test(a, y,
+        method = input$req, conf.level = as.numeric(input$ci),
+        alternative = input$alt, exact = FALSE
+      )
[…]
