+++ inst/comp_mean/report.Rmd
@@ -23,7 +23,7 @@ library(kableExtra)
 csvfile <- reactive({
     csvfile <- input$file1
     if (is.null(csvfile)){return(NULL)}
-    dt <- read.csv(csvfile$datapath, header=input$header, sep=",")
+    dt <- read.csv(csvfile$datapath, header=input$header, sep=",",check.names = FALSE)
     dt
   })
 ################# summary

[…]
