### 2021-07-08 e90ffa8
spelling correction in descApp
