+++ R/layout.R
@@ -1,21 +1,18 @@
-#'@title Field Layout of Designs
-#'@return Nothing
-#'@description
-#'layoutApp() function opens up an interactive shiny app which will allow
-#'the user to create field layout of Completely Randomized Design (CRD),
-#'Randomized Complete Block Design (RCBD), Split-plot design, Strip-plot design
-#'and Augmented Randomized complete block design. Layout generated are
-#'random. Field layout in table format can also be prepared for recording
-#'observations from the field. Results can be downloaded in word, PDF or
-#' HTML format
+#' @title Field Layout of Designs
+#' @return Nothing
+#' @description
+#' layoutApp() function opens up an interactive shiny app which will allow
+#' the user to create field layout of Completely Randomized Design (CRD),
+#' Randomized Complete Block Design (RCBD), Split-plot design, Strip-plot design
+#' and Augmented Randomized complete block design. Layout generated are
+#' random. Field layout in table format can also be prepared for recording
+#' observations from the field. Results can be downloaded in HTML format
 #'
 #' @details
-#'This app uses \code{design.crd}, \code{design.rcbd}, \code{design.dau},
-#'\code{design.strip}, \code{design.split} functions of package
-#'\code{agricolae} to generate random layout of designs. Field layout
-#'were plotted using \code{desplot} function in \code{desplot} package.
-#'To download the results in pdf one may use
-#'\code{tinytex::install_tinytex()} before using the app.
+#' This app uses \code{design.crd}, \code{design.rcbd}, \code{design.dau},
+#' \code{design.strip}, \code{design.split} functions of package
+#' \code{agricolae} to generate random layout of designs. Field layout
+#' were plotted using \code{desplot} function in \code{desplot} package.
 #'
 #' @keywords
 #' Completely Randomized Design
@@ -32,42 +29,35 @@
 #' @export
 #'
 #' @examples
-#'if(interactive()){
-#'layoutApp()
-#'}
+#' if (interactive()) {
+#'   layoutApp()
+#' }
+#' @references
 #'
 #'
-#'@references
+#' \insertRef{R_2021}{grapesAgri1}
 #'
+#' \insertRef{shiny_2021}{grapesAgri1}
 #'
-#'\insertRef{R_2021}{grapesAgri1}
+#' \insertRef{sw_2021}{grapesAgri1}
 #'
-#'\insertRef{shiny_2021}{grapesAgri1}
+#' \insertRef{shinycss_2020}{grapesAgri1}
 #'
-#'\insertRef{sw_2021}{grapesAgri1}
+#' \insertRef{dplyr_2021}{grapesAgri1}
 #'
-#'\insertRef{shinycss_2020}{grapesAgri1}
+#' \insertRef{agricolae_2020}{grapesAgri1}
 #'
-#'\insertRef{dplyr_2021}{grapesAgri1}
+#' \insertRef{desplot_2020}{grapesAgri1}
 #'
-#'\insertRef{agricolae_2020}{grapesAgri1}
+#' \insertRef{magi_2020}{grapesAgri1}
 #'
-#'\insertRef{desplot_2020}{grapesAgri1}
+#' \insertRef{Yihui_Xie_2021}{grapesAgri1}
 #'
-#'\insertRef{magi_2020}{grapesAgri1}
+#' \insertRef{gupta1985statistical}{grapesAgri1}
 #'
-#'\insertRef{Yihui_Xie_2021}{grapesAgri1}
-#'
-#'\insertRef{gupta1985statistical}{grapesAgri1}
-#'
-#'\insertRef{das1979design}{grapesAgri1}
+#' \insertRef{das1979design}{grapesAgri1}
 
-layoutApp<- function() {
[…]
