+++ inst/layout/tests/shinytest/mytest.R
@@ -1,0 +1,27 @@
+app <- ShinyDriver$new("../../", loadTimeout = 1e+05)
+app$snapshotInit("mytest")
+
+app$snapshot()
+app$setInputs(trt_crd = 3,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_crd = 5,allowInputNoBinding_ = TRUE)
+app$setInputs(submit1 = "click",allowInputNoBinding_ = TRUE)
+app$setInputs(design = "rbd",allowInputNoBinding_ = TRUE)
+app$setInputs(trt_rbd = 3,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_rbd = 7,allowInputNoBinding_ = TRUE)
+app$setInputs(submit2 = "click",allowInputNoBinding_ = TRUE)
+app$setInputs(design = "split",allowInputNoBinding_ = TRUE)
+app$setInputs(maintrt_split = 5,allowInputNoBinding_ = TRUE)
+app$setInputs(subtrt_split = 3,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_split = 3,allowInputNoBinding_ = TRUE)
+app$setInputs(submit4 = "click",allowInputNoBinding_ = TRUE)
+app$setInputs(design = "aug",allowInputNoBinding_ = TRUE)
+app$setInputs(check_aug = 2,allowInputNoBinding_ = TRUE)
+app$setInputs(trt_aug = 30,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_aug = 1,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_aug = 13,allowInputNoBinding_ = TRUE)
+app$setInputs(submit3 = "click",allowInputNoBinding_ = TRUE)
+app$setInputs(design = "strip",allowInputNoBinding_ = TRUE)
+app$setInputs(maintrt_strip = 4,allowInputNoBinding_ = TRUE)
+app$setInputs(subtrt_strip = 5,allowInputNoBinding_ = TRUE)
+app$setInputs(rep_strip = 2,allowInputNoBinding_ = TRUE)
+app$setInputs(submit5 = "click",allowInputNoBinding_ = TRUE)
