+++ inst/layout/app.R
@@ -9,56 +9,53 @@ library(shinycssloaders)
 library(knitr)
 
 ############################## Ui
- ui <- fluidPage(
+ui = fluidPage(
   setBackgroundColor(
     color = c("#fff291", "#ffffff"),
     gradient = "radial",
     direction = c("bottom", "right")
   ),
-  titlePanel(tags$div(tags$b('Field Layout of experiments'),style="color:#000000")),
-
+  titlePanel(tags$div(tags$b("Field Layout of experiments"), style = "color:#000000")),
   sidebarPanel(
-    selectInput('design', tags$div(tags$b('Select the design for your experiment'),
-                                   style="color:#d17b02"),
-                c("CRD"= 'crd',
-                  "RBD"='rbd',
-                  "Split Plot"='split',
-                  "Augmented RCBD"='aug',
-                  "Strip Plot"='strip'
-
-                )
-                ,'crd'),
+    selectInput(
+      "design", tags$div(tags$b("Select the design for your experiment"),
+        style = "color:#d17b02"
+      ),
+      c(
+        "CRD" = "crd",
+        "RBD" = "rbd",
+        "Split Plot" = "split",
+        "Augmented RCBD" = "aug",
+        "Strip Plot" = "strip"
+      ),
+      "crd"
+    ),
     h5("Note: consult a statistician before selecting proper design for your experiment"),
 
     ##
-    uiOutput('var_crd'),
-    uiOutput('var_rbd'),
-    uiOutput('var_aug'),
-    uiOutput('var_split'),
-    uiOutput('var_strip'),
-
+    uiOutput("var_crd"),
+    uiOutput("var_rbd"),
+    uiOutput("var_aug"),
+    uiOutput("var_split"),
+    uiOutput("var_strip"),
     tags$br(),
-
-    htmlOutput('text_crd1'),
-    htmlOutput('text_rbd1'),
-    htmlOutput('text_aug1'),
-    htmlOutput('text_split1'),
-    htmlOutput('text_strip1'),
-
-
-    uiOutput('var_crd1'),
-    uiOutput('var_rbd1'),
-    uiOutput('var_aug1'),
-    uiOutput('var_split1'),
-    uiOutput('var_strip1'),
-
+    htmlOutput("text_crd1"),
+    htmlOutput("text_rbd1"),
+    htmlOutput("text_aug1"),
+    htmlOutput("text_split1"),
+    htmlOutput("text_strip1"),
+    uiOutput("var_crd1"),
+    uiOutput("var_rbd1"),
+    uiOutput("var_aug1"),
+    uiOutput("var_split1"),
+    uiOutput("var_strip1"),
     tags$br(),
     h5(" If the image is not clear you can generate a table format by clicking below"),
-    uiOutput('tab_crd'),
-    uiOutput('tab_rbd'),
-    uiOutput('tab_aug'),
-    uiOutput('tab_split'),
-    uiOutput('tab_strip'),
+    uiOutput("tab_crd"),
+    uiOutput("tab_rbd"),
+    uiOutput("tab_aug"),
+    uiOutput("tab_split"),
+    uiOutput("tab_strip"),
 
     ##
 
@@ -67,7 +64,6 @@ library(knitr)
 
     ##
     tags$br(),
-
     h5(
       tags$div(
         tags$br(),
@@ -89,1216 +85,1608 @@ library(knitr)
         tags$b("MSc, Agricultural Statistics"),
         tags$br(),
         tags$br(),
-        "post your queries at: pratheesh.pg@kau.in"
-        ,style="color:#343aeb")
+        "post your queries at: pratheesh.pg@kau.in",
+        style = "color:#343aeb"
+      )
     )
-
   ),
   mainPanel(
-
[…]

[…]
