+++ .github/workflows/R-CMD-check.yaml
@@ -1,144 +1,85 @@
-name: R-CMD-check
+# For help debugging build failures open an issue on the RStudio community with the 'github-actions' tag.
+# https://community.rstudio.com/new-topic?category=Package%20development&tags=github-actions
 on:
   push:
-  branches:
-  - master
-pull_request:
-  branches:
-  - master
+    branches:
+      - main
+      - master
+  pull_request:
+    branches:
+      - main
+      - master
 
+name: R-CMD-check
 
 jobs:
-
   R-CMD-check:
-  runs-on: ${{ matrix.config.os }}
+    runs-on: ${{ matrix.config.os }}
 
-name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
 
-strategy:
-  fail-fast: false
-matrix:
-  config:
-  - {os: macOS-latest,   r: 'devel'}
-- {os: macOS-latest,   r: '4.0'}
-- {os: windows-latest, r: '4.0'}
-- {os: ubuntu-16.04,   r: '4.0', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
-- {os: ubuntu-16.04,   r: '3.6', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
-- {os: ubuntu-16.04,   r: '3.5', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
-- {os: ubuntu-16.04,   r: '3.4', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
-- {os: ubuntu-16.04,   r: '3.3', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
+    strategy:
+      fail-fast: false
+      matrix:
+        config:
+          - {os: windows-latest, r: 'release'}
+          - {os: macOS-latest, r: 'release'}
+          - {os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: ubuntu-20.04,   r: 'devel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest", http-user-agent: "R/4.1.0 (ubuntu-20.04) R (4.1.0 x86_64-pc-linux-gnu x86_64 linux-gnu) on GitHub Actions" }
 
-env:
-  _R_CHECK_FORCE_SUGGESTS_: false
-R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-RSPM: ${{ matrix.config.rspm }}
-GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      RSPM: ${{ matrix.config.rspm }}
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
-steps:
-  # https://github.com/actions/checkout/issues/135
-  - name: Set git to use LF
-if: runner.os == 'Windows'
-run: |
-  git config --system core.autocrlf false
-git config --system core.eol lf
-- uses: actions/checkout@v2
+    steps:
+      - uses: actions/checkout@v2
 
-- uses: r-lib/actions/setup-r@master
-id: install-r
-with:
-  r-version: ${{ matrix.config.r }}
+      - uses: r-lib/actions/setup-r@v1
+        with:
+          r-version: ${{ matrix.config.r }}
 
-- uses: r-lib/actions/setup-pandoc@master
+      - uses: r-lib/actions/setup-pandoc@v1
 
[…]
