+++ inst/RBD/app.R
@@ -91,7 +91,7 @@ ui <- fluidPage(
                          uiOutput('varboxplot'),
                          tags$br(),
                          uiOutput('start_note3'),
-                         plotOutput('boxplot')%>% withSpinner(color="#0dc5c1"),
+                         plotOutput('boxplot'),
                          tags$br(),
                          tags$br(),
                          uiOutput('image_down'),#image to download
@@ -150,6 +150,21 @@ ui <- fluidPage(
       if(is.null(input$file1$datapath)){return()}
       if(is.null(input$submit)){return()}
       if(input$submit > 0){
+        validate(
+          need(input$treatment != input$yield, "Warning 1: Both input variables selected (Treatment and response) are same. Choose Treatment and response correctly for meaningful result")
+        )
+        validate(
+          need(input$treatment != input$Replication, "Warning 1: Both input variables selected (Treatment and Replication) are same. Choose Treatment and Replication correctly for meaningful result")
+        )
+        validate(
+          need(input$yield != input$Replication, "Warning 1: Both input variables selected (Replication and response) are same. Choose Replication and response correctly for meaningful result")
+        )
+        validate(
+          need(input$trt!=0, "Please enter number of treatments")
+        )
+        validate(
+          need(input$rep!=0, "Please enter number of replications")
+        )
         input$reload
         Sys.sleep(2)
         d=as.data.frame(csvfile())
@@ -175,6 +190,21 @@ ui <- fluidPage(
       if(is.null(input$file1$datapath)){return()}
       if(is.null(input$submit)){return()}
       if(input$submit > 0){
+        validate(
+          need(input$treatment != input$yield, "")
+        )
+        validate(
+          need(input$treatment != input$Replication, "")
+        )
+        validate(
+          need(input$yield != input$Replication, "")
+        )
+        validate(
+          need(input$trt!=0, "")
+        )
+        validate(
+          need(input$rep!=0, "")
+        )
         d=as.data.frame(csvfile())
         r=as.numeric(input$rep)
         t=as.numeric(input$trt)
@@ -197,6 +227,21 @@ ui <- fluidPage(
       if(is.null(input$file1$datapath)){return()}
       if(is.null(input$submit)){return()}
       if(input$submit > 0){
+        validate(
+          need(input$treatment != input$yield, "")
+        )
+        validate(
+          need(input$treatment != input$Replication, "")
+        )
+        validate(
+          need(input$yield != input$Replication, "")
+        )
+        validate(
+          need(input$trt!=0, "")
+        )
+        validate(
+          need(input$rep!=0, "")
+        )
         d=as.data.frame(csvfile())
         r=as.numeric(input$rep)
         t=as.numeric(input$trt)
@@ -226,6 +271,21 @@ ui <- fluidPage(
       if(is.null(input$file1$datapath)){return()}
       if(is.null(input$submit)){return()}
       if(input$submit > 0){
+        validate(
+          need(input$treatment != input$yield, "")
+        )
+        validate(
+          need(input$treatment != input$Replication, "")
+        )
+        validate(
+          need(input$yield != input$Replication, "")
+        )
+        validate(
+          need(input$trt!=0, "")
+        )
+        validate(
+          need(input$rep!=0, "")
+        )
         d=as.data.frame(csvfile())
         r=as.numeric(input$rep)
         t=as.numeric(input$trt)
@@ -250,6 +310,21 @@ ui <- fluidPage(
       if(is.null(input$file1$datapath)){return()}
       if(is.null(input$submit)){return()}
       if(input$submit > 0){
+        validate(
+          need(input$treatment != input$yield, "")
+        )
+        validate(
+          need(input$treatment != input$Replication, "")
+        )
+        validate(
+          need(input$yield != input$Replication, "")
+        )
+        validate(
[…]
