+++ DESCRIPTION
@@ -30,7 +30,6 @@ BugReports: https://github.com/pratheesh3780/grapesAgri1/issues
 Imports: 
         shiny(>= 1.6.0),
         shinyWidgets(>= 0.6.0),
-        shinycssloaders (>= 1.0.0),
         rmarkdown(>= 2.7),
         knitr(>= 1.31),
         kableExtra(>= 1.3.4),
