### 2021-07-09 pratheesh3780/grapesAgri1#1 @pratheesh3780
@adithirgis
All suggestions were made
1. All apps updated
2.  Download formatting issue in word was due to a bug in KableExtra. Now html option is only there
3. Download image issues rectified
4. Package version added in description
5. code style and made uniform
6. Testing added for all apps
7. Validation added for unintended user inputs
8. reading materials added in Readme
9. Suggestions in paper.md incorporated

😰

😌
Cheers!
