+++ paper.bib
@@ -1,4 +1,4 @@
-@Manual{Dominic_Comtois_2021,
+@manual{Dominic_Comtois_2021,
     title = {summarytools: Tools to Quickly and Neatly Summarize Data},
     author = {Dominic Comtois},
     year = {2021},
@@ -6,7 +6,7 @@
     url = {https://CRAN.R-project.org/package=summarytools},
   }
 
-@Manual{Hao_zhu_2021,
+@manual{Hao_zhu_2021,
     title = {kableExtra: Construct Complex Table with 'kable' and Pipe Syntax},
     author = {Hao Zhu},
     year = {2021},
@@ -14,16 +14,15 @@
     url = {https://CRAN.R-project.org/package=kableExtra},
     }
 
-@Manual{Yihui_Xie_2021,
-    title = {knitr: A General-Purpose Package for Dynamic Report
-      Generation in R},
+@manual{Yihui_Xie_2021,
+    title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
     author = {Yihui Xie},
     year = {2021},
     note = {R package version 1.33},
     url = {https://yihui.org/knitr/},
     }
 
-  @Manual{R_2021,
+  @manual{R_2021,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
@@ -32,7 +31,7 @@
     url = {https://www.R-project.org/},
   }
 
-   @Manual{shiny_2021,
+   @manual{shiny_2021,
     title = {shiny: Web Application Framework for R},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
@@ -40,7 +39,7 @@
     url = {https://CRAN.R-project.org/package=shiny},
   }
 
-   @Manual{sw_2021,
+   @manual{sw_2021,
     title = {shinyWidgets: Custom Inputs Widgets for Shiny},
     author = {Victor Perrier and Fanny Meyer and David Granjon},
     year = {2021},
@@ -48,7 +47,7 @@
     url = {https://CRAN.R-project.org/package=shinyWidgets},
   }
 
-  @Manual{dplyr_2021,
+  @manual{dplyr_2021,
     title = {dplyr: A Grammar of Data Manipulation},
     author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
     year = {2021},
@@ -56,7 +55,7 @@
     url = {https://CRAN.R-project.org/package=dplyr},
   }
 
-  @Manual{ggpubr_2020,
+  @manual{ggpubr_2020,
     title = {ggpubr: 'ggplot2' Based Publication Ready Plots},
     author = {Alboukadel Kassambara},
     year = {2020},
@@ -64,7 +63,7 @@
     url = {https://CRAN.R-project.org/package=ggpubr},
   }
 
-  @Manual{past_2018,
+  @manual{past_2018,
     title = {pastecs: Package for Analysis of Space-Time Ecological Series},
     author = {Philippe Grosjean and Frederic Ibanez},
     year = {2018},
@@ -72,7 +71,7 @@
     url = {https://CRAN.R-project.org/package=pastecs},
   }
 
-   @Manual{magi_2020,
+   @manual{magi_2020,
     title = {magrittr: A Forward-Pipe Operator for R},
     author = {Stefan Milton Bache and Hadley Wickham},
     year = {2020},
@@ -80,7 +79,7 @@
     url = {https://CRAN.R-project.org/package=magrittr},
   }
 
[…]
