### 2021-07-09 pratheesh3780/grapesAgri1#1 @adithirgis
Hello!

Can you please add the points corresponding to the list above, first issue I created. I see that you have almost completed the changes I asked except adding the reference.
Once you add the corresponding response to the right number. I will proceed with the reviewing.

Thanks a lot.
