+++ joss/paper.bib
@@ -1,4 +1,189 @@
-@article{reshape2,
+@manual{Dominic_Comtois_2021,
+  title = {summarytools: Tools to Quickly and Neatly Summarize Data},
+  author = {Dominic Comtois},
+  year = {2021},
+  note = {R package version 0.9.9},
+  url = {https://CRAN.R-project.org/package=summarytools},
+}
+
+@manual{Hao_zhu_2021,
+  title = {kableExtra: Construct Complex Table with 'kable' and Pipe Syntax},
+  author = {Hao Zhu},
+  year = {2021},
+  note = {R package version 1.3.4},
+  url = {https://CRAN.R-project.org/package=kableExtra},
+}
+
+@manual{Yihui_Xie_2021,
+  title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
+  author = {Yihui Xie},
+  year = {2021},
+  note = {R package version 1.33},
+  url = {https://yihui.org/knitr/},
+}
+
+@manual{R_2021,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2021},
+  url = {https://www.R-project.org/},
+}
+
+@manual{shiny_2021,
+  title = {shiny: Web Application Framework for R},
+  author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+  year = {2021},
+  note = {R package version 1.6.0},
+  url = {https://CRAN.R-project.org/package=shiny},
+}
+
+@manual{sw_2021,
+  title = {shinyWidgets: Custom Inputs Widgets for Shiny},
+  author = {Victor Perrier and Fanny Meyer and David Granjon},
+  year = {2021},
+  note = {R package version 0.6.0},
+  url = {https://CRAN.R-project.org/package=shinyWidgets},
+}
+
+@manual{dplyr_2021,
+  title = {dplyr: A Grammar of Data Manipulation},
+  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
+  year = {2021},
+  note = {R package version 1.0.4},
+  url = {https://CRAN.R-project.org/package=dplyr},
+}
+
+@manual{ggpubr_2020,
+  title = {ggpubr: 'ggplot2' Based Publication Ready Plots},
+  author = {Alboukadel Kassambara},
+  year = {2020},
+  note = {R package version 0.4.0},
+  url = {https://CRAN.R-project.org/package=ggpubr},
+}
+
+@manual{past_2018,
+  title = {pastecs: Package for Analysis of Space-Time Ecological Series},
+  author = {Philippe Grosjean and Frederic Ibanez},
+  year = {2018},
+  note = {R package version 1.3.21},
+  url = {https://CRAN.R-project.org/package=pastecs},
+}
+
+@manual{magi_2020,
+  title = {magrittr: A Forward-Pipe Operator for R},
+  author = {Stefan Milton Bache and Hadley Wickham},
+  year = {2020},
+  note = {R package version 2.0.1},
+  url = {https://CRAN.R-project.org/package=magrittr},
+}
+
+@manual{gridG_2020,
+  title = {gridGraphics: Redraw Base Graphics Using 'grid' Graphics},
+  author = {Paul Murrell and Zhijian Wen},
+  year = {2020},
+  note = {R package version 0.5-1},
+  url = {https://CRAN.R-project.org/package=gridGraphics},
+}
+
+@book{gupta1985statistical,
+  title={Statistical methods},
[…]
