+++ joss/paper.md
@@ -1,101 +1,71 @@
 ---
-title: 'grapesAgri1: Collection of Shiny Apps for Data Analysis in Agriculture'
+  title: 'Gala: A Python package for galactic dynamics'
 tags:
-  - R
-  - Data analysis in Agriculture
-  - shiny app
-  - Design of experiments
-  - Compare means
-  - Field Layout
-  - Correlation Analysis
-  - Descriptive Statistics
-  
+  - Python
+- astronomy
+- dynamics
+- galactic dynamics
+- milky way
 authors:
-  - name: Pratheesh P. Gopinath
-    orcid: 0000-0003-3290-0436
-    affiliation: "1"
-  - name: Rajender Parsad
-    affiliation: "2"
-  - name: Brigit Joseph
-    affiliation: "1"
-  - name: Adarsh V. S.
-    affiliation: "3"
-    
+  - name: Adrian M. Price-Whelan^[co-first author] # note this makes a footnote saying 'co-first author'
+orcid: 0000-0003-0872-7098
+affiliation: "1, 2" # (Multiple affiliations must be quoted)
+- name: Author Without ORCID^[co-first author] # note this makes a footnote saying 'co-first author'
+affiliation: 2
+- name: Author with no affiliation^[corresponding author]
+affiliation: 3
 affiliations:
- - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural University, Kerala, India.
-   index: 1
- - name: ICAR-Indian Agricultural Statistics Research Institute, New Delhi, India.
-   index: 2
- - name: Department of Agricultural Statistics, BCKV, West Bengal, India
-   index: 3
-   
-date: June 11, 2021 
+  - name: Lyman Spitzer, Jr. Fellow, Princeton University
+index: 1
+- name: Institution Name
+index: 2
+- name: Independent Researcher
+index: 3
+date: 13 August 2017
 bibliography: paper.bib
 
 ---
-
+  
 # Summary
-
-<p>
-Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR (Indian Council of Agricultural Research) for data analysis, though it is a highly advanced statistical analysis platform, its complexity hold back students and researchers from using it. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
-</p>
-<p>
-R, which is an open source software, provides an excellent platform for data analysis. This powerful computation platform is not much utilised by researchers in the agricultural field. Programming and computational knowledge hinders agricultural researchers from using R for analysis. Also, researchers in Agriculture prefer a graphical user interface.
-</p>
-<p>
-Efforts should be made to develop a high quality, reliable open-source platform with a simple interactive user interface for data analysis in agriculture. In this light utilising the features of shiny package in R, we have developed a collection of shiny apps for agricultural research called `grapesAgri1` (General R-shiny based Analysis Platform Empowered by Statistics for data analysis in Agriculture-part1). It is hosted on the web (www.kaugrapes.com), downloadable as a standalone application and now we have realised it in to an R package.
-</p>
-<p>
[…]
