+++ joss/paper.md
@@ -1,71 +1,101 @@
 ---
-  title: 'Gala: A Python package for galactic dynamics'
+  title: 'grapesAgri1: Collection of Shiny Apps for Data Analysis in Agriculture'
 tags:
-  - Python
-- astronomy
-- dynamics
-- galactic dynamics
-- milky way
+  - R
+- Data analysis in Agriculture
+- shiny app
+- Design of experiments
+- Compare means
+- Field Layout
+- Correlation Analysis
+- Descriptive Statistics
+
 authors:
-  - name: Adrian M. Price-Whelan^[co-first author] # note this makes a footnote saying 'co-first author'
-orcid: 0000-0003-0872-7098
-affiliation: "1, 2" # (Multiple affiliations must be quoted)
-- name: Author Without ORCID^[co-first author] # note this makes a footnote saying 'co-first author'
-affiliation: 2
-- name: Author with no affiliation^[corresponding author]
-affiliation: 3
+  - name: Pratheesh P. Gopinath
+orcid: 0000-0003-3290-0436
+affiliation: "1"
+- name: Rajender Parsad
+affiliation: "2"
+- name: Brigit Joseph
+affiliation: "1"
+- name: Adarsh V. S.
+affiliation: "3"
+
 affiliations:
-  - name: Lyman Spitzer, Jr. Fellow, Princeton University
+  - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural University, Kerala, India.
 index: 1
-- name: Institution Name
+- name: ICAR-Indian Agricultural Statistics Research Institute, New Delhi, India.
 index: 2
-- name: Independent Researcher
+- name: Department of Agricultural Statistics, BCKV, West Bengal, India
 index: 3
-date: 13 August 2017
+
+date: June 11, 2021 
 bibliography: paper.bib
 
 ---
   
-# Summary
+  # Summary
   
-  The forces on stars, galaxies, and dark matter under external gravitational
-fields lead to the dynamical evolution of structures in the universe. The orbits
-of these bodies are therefore key to understanding the formation, history, and
-future state of galaxies. The field of "galactic dynamics," which aims to model
-the gravitating components of galaxies to study their structure and evolution,
-is now well-established, commonly taught, and frequently used in astronomy.
-Aside from toy problems and demonstrations, the majority of problems require
-efficient numerical tools, many of which require the same base code (e.g., for
-                                                                     performing numerical orbit integration).
+  <p>
+  Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR (Indian Council of Agricultural Research) for data analysis, though it is a highly advanced statistical analysis platform, its complexity hold back students and researchers from using it. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
+</p>
+<p>
+R, which is an open source software, provides an excellent platform for data analysis. This powerful computation platform is not much utilised by researchers in the agricultural field. Programming and computational knowledge hinders agricultural researchers from using R for analysis. Also, researchers in Agriculture prefer a graphical user interface.
+</p>
+<p>
[…]
