+++ joss/paper.bib
@@ -1,203 +1,141 @@
-@manual{Dominic_Comtois_2021,
-  title = {summarytools: Tools to Quickly and Neatly Summarize Data},
-  author = {Dominic Comtois},
-  year = {2021},
-  note = {R package version 0.9.9},
-  url = {https://CRAN.R-project.org/package=summarytools},
+@article{Bollerslev1986,
+author = {Bollerslev, Tim},
+doi = {10.1016/0304-4076(86)90063-1},
+journal = {Journal of Econometrics},
+number = {3},
+pages = {307--327},
+title = {{Generalized autoregressive conditional heteroskedasticity}},
+volume = {31},
+year = {1986}
 }
 
-@manual{Hao_zhu_2021,
-  title = {kableExtra: Construct Complex Table with 'kable' and Pipe Syntax},
-  author = {Hao Zhu},
-  year = {2021},
-  note = {R package version 1.3.4},
-  url = {https://CRAN.R-project.org/package=kableExtra},
+@article{Bollerslev1990,
+author = {Bollerslev, Tim},
+doi = {10.2307/2109358},
+journal = {The Review of Economics and Statistics},
+number = {3},
+pages = {498----505},
+title = {{Modelling the Coherence in Short-Run Nominal Exchange Rates: A Multivariate Generalized Arch Model}},
+volume = {72},
+year = {1990}
 }
 
-@manual{Yihui_Xie_2021,
-  title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
-  author = {Yihui Xie},
-  year = {2021},
-  note = {R package version 1.33},
-  url = {https://yihui.org/knitr/},
+@article{Buerkner2020,
+author = {B{\"{u}}rkner, Paul-Christian and Gabry, Jonah and Vehtari, Aki},
+journal = {Journal of Statistical Computation and Simulation},
+pages = {1--25},
+title = {{Approximate leave-future-out cross-validation for Bayesian time series models}},
+doi = {10.1080/00949655.2020.1783262},
+year = {2020}
 }
 
-@manual{R_2021,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2021},
-  url = {https://www.R-project.org/},
+
+
+@article{Carnero2014,
+author = {Carnero, M. Angeles and Eratalay, M. Hakan},
+doi = {10.1515/snde-2012-0065},
+journal = {Studies in Nonlinear Dynamics {\&} Econometrics},
+keywords = {C32,financial markets,volatility spillovers},
+number = {3},
+pages = {339--365},
+title = {{Estimating VAR-MGARCH models in multiple steps}},
+volume = {18},
+year = {2014}
 }
 
-@manual{shiny_2021,
-  title = {shiny: Web Application Framework for R},
-  author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
-  year = {2021},
-  note = {R package version 1.6.0},
-  url = {https://CRAN.R-project.org/package=shiny},
+@article{DeAlmeida2018,
+author = {de Almeida, Daniel and Hotta, Luiz K. and Ruiz, Esther},
+doi = {10.1016/J.IJFORECAST.2017.08.003},
+journal = {International Journal of Forecasting},
+number = {1},
+pages = {45--63},
+title = {{MGARCH models: Trade-off between feasibility and flexibility}},
+volume = {34},
+year = {2018}
 }
 
[…]
