### 2021-07-09 pratheesh3780/grapesAgri1#1 @adithirgis
Hi!

1. I think because there are no references inside the paper. Can I please again request you to go through the example in the submission guidelines of JOSS.

eg - where `[@astropy]` is a reference
`Astropy package [@astropy]`

And also requesting you to add the numbers and the corresponding response to it as per my first list (since its difficult to keep track), so that I can check them off.

2. Also for this line -

`Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.`

If this sounds correct change it to -

`Some of the web applications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community are user friendly but these applications don't provide options to generate plots and graphs.`

3. Please also do mention what package you used to style your code.

4. And also run `devtools::check()`, some tests seem to fail.

Thanks!
Adithi
