+++ joss/paper.md
@@ -1,43 +1,43 @@
 ---
-  title: 'grapesAgri1: Collection of Shiny Apps for Data Analysis in Agriculture'
+title: 'grapesAgri1: Collection of Shiny Apps for Data Analysis in Agriculture'
 tags:
   - R
-- Data analysis in Agriculture
-- shiny app
-- Design of experiments
-- Compare means
-- Field Layout
-- Correlation Analysis
-- Descriptive Statistics
-
+  - Data analysis in Agriculture
+  - shiny app
+  - Design of experiments
+  - Compare means
+  - Field Layout
+  - Correlation Analysis
+  - Descriptive Statistics
+  
 authors:
   - name: Pratheesh P. Gopinath
-orcid: 0000-0003-3290-0436
-affiliation: "1"
-- name: Rajender Parsad
-affiliation: "2"
-- name: Brigit Joseph
-affiliation: "1"
-- name: Adarsh V. S.
-affiliation: "3"
-
+    orcid: 0000-0003-3290-0436
+    affiliation: "1"
+  - name: Rajender Parsad
+    affiliation: "2"
+  - name: Brigit Joseph
+    affiliation: "1"
+  - name: Adarsh V. S.
+    affiliation: "3"
+    
 affiliations:
-  - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural University, Kerala, India.
-index: 1
-- name: ICAR-Indian Agricultural Statistics Research Institute, New Delhi, India.
-index: 2
-- name: Department of Agricultural Statistics, BCKV, West Bengal, India
-index: 3
-
+ - name: Department of Agricultural Statistics, College of Agriculture, Vellayani, Kerala Agricultural University, Kerala, India.
+   index: 1
+ - name: ICAR-Indian Agricultural Statistics Research Institute, New Delhi, India.
+   index: 2
+ - name: Department of Agricultural Statistics, BCKV, West Bengal, India
+   index: 3
+   
 date: June 11, 2021 
 bibliography: paper.bib
 
 ---
-  
-  # Summary
-  
-  <p>
-  Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR (Indian Council of Agricultural Research) for data analysis, though it is a highly advanced statistical analysis platform, its complexity hold back students and researchers from using it. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
+
+# Summary
+
+<p>
+Agricultural experiments demand a wide range of statistical tools for analysis, which includes from Exploratory analysis to Design of experiments and Statistical genetics. It is a challenge for the scientists and the students to find a suitable platform for data analysis and to eventually publish the research outputs in quality journals. Most of the software available for data analysis are proprietary or lack a simple user interface, for example SAS is available in ICAR (Indian Council of Agricultural Research) for data analysis, though it is a highly advanced statistical analysis platform, its complexity hold back students and researchers from using it. Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.
 </p>
 <p>
[…]
