+++ joss/paper.bib
@@ -3,7 +3,7 @@
   author = {Dominic Comtois},
   year = {2021},
   note = {R package version 0.9.9},
-  url = {https://CRAN.R-project.org/package=summarytools},
+  url = {https://CRAN.R-project.org/package=summarytools}
 }
 
 @manual{Hao_zhu_2021,
@@ -11,7 +11,7 @@
   author = {Hao Zhu},
   year = {2021},
   note = {R package version 1.3.4},
-  url = {https://CRAN.R-project.org/package=kableExtra},
+  url = {https://CRAN.R-project.org/package=kableExtra}
 }
 
 @manual{Yihui_Xie_2021,
@@ -19,7 +19,7 @@
   author = {Yihui Xie},
   year = {2021},
   note = {R package version 1.33},
-  url = {https://yihui.org/knitr/},
+  url = {https://yihui.org/knitr/}
 }
 
 @manual{R_2021,
@@ -28,7 +28,7 @@
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2021},
-  url = {https://www.R-project.org/},
+  url = {https://www.R-project.org/}
 }
 
 @manual{shiny_2021,
@@ -36,7 +36,7 @@
   author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
   year = {2021},
   note = {R package version 1.6.0},
-  url = {https://CRAN.R-project.org/package=shiny},
+  url = {https://CRAN.R-project.org/package=shiny}
 }
 
 @manual{sw_2021,
@@ -44,7 +44,7 @@
   author = {Victor Perrier and Fanny Meyer and David Granjon},
   year = {2021},
   note = {R package version 0.6.0},
-  url = {https://CRAN.R-project.org/package=shinyWidgets},
+  url = {https://CRAN.R-project.org/package=shinyWidgets}
 }
 
 @manual{dplyr_2021,
@@ -52,7 +52,7 @@
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
   year = {2021},
   note = {R package version 1.0.4},
-  url = {https://CRAN.R-project.org/package=dplyr},
+  url = {https://CRAN.R-project.org/package=dplyr}
 }
 
 @manual{ggpubr_2020,
@@ -60,7 +60,7 @@
   author = {Alboukadel Kassambara},
   year = {2020},
   note = {R package version 0.4.0},
-  url = {https://CRAN.R-project.org/package=ggpubr},
+  url = {https://CRAN.R-project.org/package=ggpubr}
 }
 
 @manual{past_2018,
@@ -68,7 +68,7 @@
   author = {Philippe Grosjean and Frederic Ibanez},
   year = {2018},
   note = {R package version 1.3.21},
-  url = {https://CRAN.R-project.org/package=pastecs},
+  url = {https://CRAN.R-project.org/package=pastecs}
 }
 
 @manual{magi_2020,
@@ -76,7 +76,7 @@
   author = {Stefan Milton Bache and Hadley Wickham},
   year = {2020},
   note = {R package version 2.0.1},
-  url = {https://CRAN.R-project.org/package=magrittr},
+  url = {https://CRAN.R-project.org/package=magrittr}
 }
 
 @manual{gridG_2020,
@@ -84,7 +84,7 @@
   author = {Paul Murrell and Zhijian Wen},
   year = {2020},
   note = {R package version 0.5-1},
-  url = {https://CRAN.R-project.org/package=gridGraphics},
+  url = {https://CRAN.R-project.org/package=gridGraphics}
 }
 
[…]
