### 2021-07-09 pratheesh3780/grapesAgri1#1 @pratheesh3780
> Thank you for submitting to JOSS, I enjoyed reviewing this package. Congratulations!
>
> Starting my review here. To avoid system errors I tried to run this package on [RStudio cloud](https://rstudio.cloud/).
>
> 1. Can you please add references in [paper.md](https://github.com/adithirgis/grapesAgri1/blob/master/paper.md) using [paper.bib](https://github.com/adithirgis/grapesAgri1/blob/master/paper.bib).

- **Sorry for not following up. I'm on to it**

> 2. Can you please mention whether `grapesAgri1` has been cited or referenced anywhere and if yes where please add it.

- **Its new but its now used by a lot of students not cited anywhere**

> 3. As mentioned in the [paper.md](https://github.com/adithirgis/grapesAgri1/blob/master/paper.md) -
>    "Scientist and research students were in trouble to find a suitable platform for their data analysis and also to publish the research output in quality journals. Most of the softwares available for data analysis are proprietary or lacks a simple user interface. Though several online platforms were available for data analysis, they are not much updated up to the current publication standards at international level."

- **I have added these suggestions in [paper.md](https://github.com/adithirgis/grapesAgri1/blob/master/paper.md)**

>    Can the authors please mention software name they have come across which are proprietary, lack simple user interface or not updated to current standards and do not provide graphs, thus providing a comparative study.

- **I have added these suggestions in [paper.md](https://github.com/adithirgis/grapesAgri1/blob/master/paper.md)**

> 4. Also in the [DESCRIPTION](https://github.com/adithirgis/grapesAgri1/blob/master/DESCRIPTION) of the package, please add the version number of the packages which are being imported, this might resolve the conflict of versions.

- **Package version added**

> 5. Styling of the code is not uniform , please refer to this guide for styling or you could also use `styler` package, sometimes `=` is used for assignment and in other places `<-` is used (recommend choosing one style and sticking with it)
>    eg -
>    `result=as.data.frame( anova(anvaTable) ); trtmeans<-out$means`

- **Styling and changes were done, "=" is used uniformly everywhere,  styler package is used**

> 6. Spacing in code is inconsistent. Also the code lines stretch very long, beyond the recommended 80 characters or 120 characters. Here is a resource which could be helpful - https://devguide.ropensci.org/building.html#code-style, http://adv-r.had.co.nz/Style.html and https://www.tidyverse.org/blog/2017/12/styler-1.0.0/.

- **Done**

> 7. Can you please point me to the tests folder of this application? I do not seem to find the test folder in the package. Please add test units for automatically testing your application. Tests for functions will also be required. Examples from JOSS - https://github.com/michaellevy/gwdegree/tree/master/tests/testthat, https://github.com/PsyChiLin/EFAshiny/tree/master/tests/testthat, https://github.com/OttaviaE/DscoreApp/tree/master/tests, https://github.com/pmcharrison/psychTestR/tree/master/tests/testthat.

- **Test where added using shinytest and testthat [here](https://github.com/pratheesh3780/grapesAgri1/tree/master/tests)**

> 8. I got three notes while checking the package using `devtools::check()`. I would suggest adding `.bib` and `.md` extensions to the `.Rbuildignore` file.  And another problem I faced is that I require `kableExtra` as an import not suggests, because I tried installing the package and it still says I need to install the `kableExtra`, `summarytools`, and for `grapesAgri1::ttApp()` it required `PairedData`.

- **Done. All these packages are now in import**

>    `checking package dependencies ... NOTE Packages suggested but not available for checking:   'kableExtra', 'summarytools', 'desplot', 'PairedData'`.

- **Resolved issue**

> 9. When I choose the car.csv and try to use `grapesAgri1::descApp()` without header it shows me this error -
>    ![image](https://user-images.githubusercontent.com/63432077/124088055-35477580-da70-11eb-99cb-a22d146ddac8.png)

- **Header should be ticked if you are using a file with header, I hope it is clear**

> 10. For correlation analysis report download in `grapesAgri1::corrApp()` function for Simple_corelation, for html version the formatting looks fine, while the word/docx version does not seem very intuitive, is there a way to format it correctly, and the pdf version is not downloaded in my windows 10 system. Error for pdf version is - `Warning in system2(..., stdout = if (use_file_stdout()) f1 else FALSE, stderr = f2) : '"pdflatex"' not found Warning: Error in : LaTeX failed to compile report.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips.`-
>
> ![image](https://user-images.githubusercontent.com/63432077/124088155-4abc9f80-da70-11eb-94e8-292f65436892.png)
> ![image](https://user-images.githubusercontent.com/63432077/124088184-514b1700-da70-11eb-8cd5-89088086dda6.png)
> ![image](https://user-images.githubusercontent.com/63432077/124088215-57d98e80-da70-11eb-9cdf-f405124fbdc9.png)
>

**- Problem cause by KableExtra as several apps are incoporated together. Removed PDF and Word download option, now only html, good to go!**

> 1. Also when I don't select any options it shows this error, is it possible to show these errors as warnings or have an option already selected -
>
> ![image](https://user-images.githubusercontent.com/63432077/124088362-7dff2e80-da70-11eb-95b4-506d596c6f78.png)
>

**- Validation added for unintended user inputs**

> 1. For this function `grapesAgri1::crdApp()` for the below selected parameters I am not able to generate the table is there some mistake I am making, please explain the data used here `unequal_replication.csv` and others to understand the functionality better -
>    ![image](https://user-images.githubusercontent.com/63432077/124089613-ab98a780-da71-11eb-9e6b-e741f4be643d.png)

- **Validation added for unintended user inputs**

> 2. It would be great if these functions - `grapesAgri1::layoutApp()` and `grapesAgri1::rbdApp()` and its application are explained in the [ReadMe](https://github.com/adithirgis/grapesAgri1#readme) file. And to add both the apps work well with the `mango data.csv`.

**- layout app generates field layout of experiments to perform experiment in field.  Additional reading materials added in Readme**

> 3. Paper suggestions - I will [pr](https://github.com/pratheesh3780/grapesAgri1/pull/2), please feel free to ignore some if you feel its not grammatically correct. Changes are in bold and strikeout.

- **All your suggestions were incorporated**

> 4. Extra suggestion, if possible the Community guidelines can be mentioned in the documentation or the [ReadMe](https://github.com/adithirgis/grapesAgri1#readme) file instead of [paper.md](https://github.com/adithirgis/grapesAgri1/blob/master/paper.md)

**- Done**
