### 2021-07-09 pratheesh3780/grapesAgri1#1 @adithirgis
Hello,

Thanks for following up soon! Really appreciate it.

I have a few things in my checklist now to end the review -

- [ ] Run `devtools::check()` and `devtools::test()`, some tests seem to fail in my system. Let me add the error I get here. This is using the latest version of the app also `devtools::check()` also gives the same error.

```r
> devtools::test()
i Loading grapesAgri1
Registered S3 method overwritten by 'pryr':
  method      from
  print.bytes Rcpp
i Testing grapesAgri1
√ |  OK F W S | Context
x |   0 1     | app-file [70.3 s]
-------------------------------------------------------------------------------------------
Failure (testdesc.R:6:3): descApp_testing
Not all shinytest scripts passed for D:/Dropbox/ILKConsultancy/grapesAgri1/inst/desc: mytest

Diff output:
==== mytest ====
diff "C:\\Users\\Dell\\AppData\\Local\\Temp\\Rtmpo35dGe\\shinytest-diff-1eec7d913a59/mytest-expected/001.json" "C:\\Users\\Dell\\AppData\\Local\\Temp\\Rtmpo35dGe\\shinytest-diff-1eec7d913a59/mytest-current/001.json"
1,68c1,68
< {
<   "input": {
<     "dataset": "iris",
<     "file1": null,
<     "header": true,
<     "req": "summary",
<     "shiny_height": 744,
<     "shiny_width": 992
<   },
<   "output": {
<     "bygroup": null,
<     "data_set": {
<       "html": "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"dataset-label\" for=\"dataset\">Choose a dataset:<\/label>\n  <div>\n    <select id=\"dataset\"><option value=\"iris\" selected>iris<\/option>\n<option value=\"pressure\">pressure<\/option>\n<option value=\"cars\">cars<\/option><\/select>\n    <script type=\"application/json\" data-for=\"dataset\" data-nonempty=\"\">{\"plugins\":[\"selectize-plugin-a11y\"]}<\/script>\n  <\/div>\n<\/div>\n<a id=\"downloadData\" class=\"btn btn-default shiny-download-link butt1\" href=\"\" target=\"_blank\" download>\n  <i class=\"fa fa-download\" role=\"presentation\" aria-label=\"download icon\"><\/i>\n  Download csv file to test\n<\/a>",
<       "deps": [
<         {
<           "name": "selectize",
<           "version": "0.12.4",
<           "src": {
<             "href": "shared/selectize"
<           },
<           "meta": null,
<           "script": [
<             "js/selectize.min.js",
<             "accessibility/js/selectize-plugin-a11y.min.js"
<           ],
<           "stylesheet": "css/selectize.bootstrap3.css",
<           "head": null,
<           "attachment": null,
<           "package": null,
<           "all_files": true
<         },
<         {
<           "name": "font-awesome",
<           "version": "5.13.0",
<           "src": {
<             "href": "font-awesome-5.13.0"
<           },
<           "meta": null,
<           "script": null,
<           "stylesheet": [
<             "css/all.min.css",
<             "css/v4-shims.min.css"
<           ],
<           "head": null,
<           "attachment": null,
<           "all_files": true
<         }
<       ]
<     },
<     "image_down": null,
<     "nort": "",
<     "note": {
<       "html": " <h4> To perform analysis using your own dataset prepare excel file in csv format by reading instruction below  <\/h4>\n<p>\n<ui>\n<li>Open a new blank excel file<\/li>\n<li>Copy and paste observations into a new sheet (use only one sheet) of a new excel file<\/li>\n<li>Observations should be pasted as columns <\/li>\n<li>Don't type or delete anything on other cells without data<\/li>\n<li>You can use any names for your columns. No space is allowed in the Column name. If space is required use underscore ‘_’ or ‘.’ full stop; for example ‘Variable name’ should be written as Variable_name or Variable.name<\/li>\n<li>Data should be arranged towards upper left corner and row above the data should not be left blank <\/li>\n<li>Don't type and delete anything on other cells without data. If so select those cells, right click and click clear contents <\/li>\n<li>Give names to all column, Don't add any unnecessary columns that is not required for analysis<\/li>\n<li>Once all these are done, your file is ready. Now save it as CSV file. <\/li>\n<li><b>Upload file by clicking browse in the app <\/b><\/li>\n<\/ui>\n<\/p>\n<h5> You can download a model data set from below and test the App  <\/h5>\n",
<       "deps": [
<
<       ]
<     },
<     "plot": null,
<     "summaryout": null,
<     "text3": null,
<     "var": null,
<     "var1": null,
<     "var2": null
<   },
<   "export": {
<
<   }
< }
---
> {
>   "input": {
>     "dataset": "iris",
>     "file1": null,
>     "header": true,
>     "req": "summary",
>     "shiny_height": 744,
>     "shiny_width": 992
>   },
>   "output": {
>     "bygroup": null,
>     "data_set": {
>       "html": "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"dataset-label\" for=\"dataset\">Choose a dataset:<\/label>\n  <div>\n    <select id=\"dataset\"><option value=\"iris\" selected>iris<\/option>\n<option value=\"pressure\">pressure<\/option>\n<option value=\"cars\">cars<\/option><\/select>\n    <script type=\"application/json\" data-for=\"dataset\" data-nonempty=\"\">{\"plugins\":[\"selectize-plugin-a11y\"]}<\/script>\n  <\/div>\n<\/div>\n<a id=\"downloadData\" class=\"btn btn-default shiny-download-link butt1\" href=\"\" target=\"_blank\" download>\n  <i class=\"fa fa-download\" role=\"presentation\" aria-label=\"download icon\"><\/i>\n  Download csv file to test\n<\/a>",
>       "deps": [
>         {
>           "name": "selectize",
>           "version": "0.12.4",
>           "src": {
>             "href": "shared/selectize"
>           },
>           "meta": null,
>           "script": [
>             "js/selectize.min.js",
>             "accessibility/js/selectize-plugin-a11y.min.js"
>           ],
>           "stylesheet": "css/selectize.bootstrap3.css",
>           "head": null,
>           "attachment": null,
>           "package": null,
>           "all_files": true
>         },
>         {
>           "name": "font-awesome",
>           "version": "5.13.0",
>           "src": {
>             "href": "font-awesome-5.13.0"
>           },
>           "meta": null,
>           "script": null,
>           "stylesheet": [
>             "css/all.min.css",
>             "css/v4-shims.min.css"
>           ],
>           "head": null,
>           "attachment": null,
>           "all_files": true
>         }
>       ]
>     },
>     "image_down": null,
>     "nort": "",
>     "note": {
>       "html": " <h4> To perform analysis using your own dataset prepare excel file in csv format by reading instruction below  <\/h4>\n<p>\n<ui>\n<li>Open a new blank excel file<\/li>\n<li>Copy and paste observations into a new sheet (use only one sheet) of a new excel file<\/li>\n<li>Observations should be pasted as columns <\/li>\n<li>Don't type or delete anything on other cells without data<\/li>\n<li>You can use any names for your columns. No space is allowed in the Column name. If space is required use underscore ‘_’ or ‘.’ full stop; for example ‘Variable name’ should be written as Variable_name or Variable.name<\/li>\n<li>Data should be arranged towards upper left corner and row above the data should not be left blank <\/li>\n<li>Don't type and delete anything on other cells without data. If so select those cells, right click and click clear contents <\/li>\n<li>Give names to all column, Don't add any unnecessary columns that is not required for analysis<\/li>\n<li>Once all these are done, your file is ready. Now save it as CSV file. <\/li>\n<li><b>Upload file by clicking browse in the app <\/b><\/li>\n<\/ui>\n<\/p>\n<h5> You can download a model data set from below and test the App  <\/h5>\n",
>       "deps": [
>
>       ]
>     },
>     "plot": null,
>     "summaryout": null,
>     "text3": null,
>     "var": null,
>     "var1": null,
>     "var2": null
>   },
>   "export": {
>
>   }
> }

If this is expected, use `snapshotUpdate('D:/Dropbox/ILKConsultancy/grapesAgri1/inst/desc')` to update
-------------------------------------------------------------------------------------------
x |   0 1     | app-file [21.4 s]
-------------------------------------------------------------------------------------------
Failure (testttest.R:6:3): ttest_testing
Not all shinytest scripts passed for D:/Dropbox/ILKConsultancy/grapesAgri1/inst/comp_mean: mytest

Diff output:
==== mytest ====
diff "C:\\Users\\Dell\\AppData\\Local\\Temp\\Rtmpo35dGe\\shinytest-diff-1eec37707b35/mytest-expected/001.json" "C:\\Users\\Dell\\AppData\\Local\\Temp\\Rtmpo35dGe\\shinytest-diff-1eec37707b35/mytest-current/001.json"
1,81c1,81
< {
<   "input": {
<     "dataset": "iris",
<     "file1": null,
<     "header": true,
<     "req1": "ttest"
<   },
<   "output": {
<     "data_set": {
<       "html": "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"dataset-label\" for=\"dataset\">Choose a dataset:<\/label>\n  <div>\n    <select id=\"dataset\"><option value=\"iris\" selected>iris<\/option>\n<option value=\"rock\">rock<\/option>\n<option value=\"pressure\">pressure<\/option><\/select>\n    <script type=\"application/json\" data-for=\"dataset\" data-nonempty=\"\">{\"plugins\":[\"selectize-plugin-a11y\"]}<\/script>\n  <\/div>\n<\/div>\n<a id=\"downloadData\" class=\"btn btn-default shiny-download-link butt1\" href=\"\" target=\"_blank\" download>\n  <i class=\"fa fa-download\" role=\"presentation\" aria-label=\"download icon\"><\/i>\n  Download csv file to test\n<\/a>",
<       "deps": [
<         {
<           "name": "selectize",
<           "version": "0.12.4",
<           "src": {
<             "href": "shared/selectize"
<           },
<           "meta": null,
<           "script": [
<             "js/selectize.min.js",
<             "accessibility/js/selectize-plugin-a11y.min.js"
<           ],
<           "stylesheet": "css/selectize.bootstrap3.css",
<           "head": null,
<           "attachment": null,
<           "package": null,
<           "all_files": true
<         },
<         {
<           "name": "font-awesome",
<           "version": "5.13.0",
<           "src": {
<             "href": "font-awesome-5.13.0"
<           },
<           "meta": null,
<           "script": null,
<           "stylesheet": [
<             "css/all.min.css",
<             "css/v4-shims.min.css"
<           ],
<           "head": null,
<           "attachment": null,
<           "all_files": true
<         }
<       ]
<     },
<     "ftest": null,
<     "image_down": null,
<     "note1": null,
<     "note2": null,
<     "note3": null,
<     "note4": null,
<     "note5": null,
<     "ottest": null,
<     "plot": null,
<     "pttest": null,
<     "slider": null,
<     "start_note": {
<       "html": " <h4> To perform analysis using your own dataset, prepare excel file in csv format by reading instruction below  <\/h4>\n<p>\n<ui>\n<li>Open a new blank excel file<\/li>\n<li>Copy and paste observations into a new sheet (use only one sheet) of a new excel file<\/li>\n<li>Groups to be compared should be pasted as columns <\/li>\n<li>Don't type or delete anything on other cells without data<\/li>\n<li>You can use any names for your columns. No space is allowed in the Column name. If space is required use underscore ‘_’ or ‘.’ full stop; for example ‘Variable name’ should be written as Variable_name or Variable.name<\/li>\n<li>Data should be arranged towards upper left corner and row above the data should not be left blank <\/li>\n<li>Type 'NA' in the cell with no observation<\/li>\n<li>Don't type and delete anything on other cells without data. If so select those cells, right click and click clear contents <\/li>\n<li>Give names to all column, Don't add any unnecessary columns that is not required for analysis<\/li>\n<li>Once all these are done, your file is ready. Now save it as CSV file. <\/li>\n<li><b>Upload file by clicking browse in the app <\/b><\/li>\n<\/ui>\n<\/p>\n<p>\n<b> Things to remember while doing t-test<\/b>\n<li>Note that, unpaired two-samples t-test can be used only under certain conditions: <\/li>\n<li>First condition: each of the two groups of samples being compared, should be normally distributed. This can be checked using Shapiro-Wilk test and other data visualization techniques [use grapes::descApp()]<\/li>\n<li>Second condition: Variances of the two groups should be equal.\nThis can be checked using Homogenity of variance test (F test) available\nwith in the app. <\/li>\n<li>If the variances of the two groups being compared are different (heteroscedasticity), it’s possible to use the Welch t test, an adaptation of Student t-test. <\/li>\n<li>Welch t test and F-test are available within the app<\/li>\n<\/p>\n<b> You can download a model data set from below and test the App  <\/b>\n",
<       "deps": [
<
<       ]
<     },
<     "summary1": null,
<     "summary2": null,
<     "summary3": null,
<     "summary4": null,
<     "summary5": null,
<     "ttest": null,
<     "var": null,
<     "var1": null,
<     "var2": null,
<     "var3": null,
<     "var4": null,
<     "var5": null,
<     "wttest": null
<   },
<   "export": {
<
<   }
< }
---
> {
>   "input": {
>     "dataset": "iris",
>     "file1": null,
>     "header": true,
>     "req1": "ttest"
>   },
>   "output": {
>     "data_set": {
>       "html": "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"dataset-label\" for=\"dataset\">Choose a dataset:<\/label>\n  <div>\n    <select id=\"dataset\"><option value=\"iris\" selected>iris<\/option>\n<option value=\"rock\">rock<\/option>\n<option value=\"pressure\">pressure<\/option><\/select>\n    <script type=\"application/json\" data-for=\"dataset\" data-nonempty=\"\">{\"plugins\":[\"selectize-plugin-a11y\"]}<\/script>\n  <\/div>\n<\/div>\n<a id=\"downloadData\" class=\"btn btn-default shiny-download-link butt1\" href=\"\" target=\"_blank\" download>\n  <i class=\"fa fa-download\" role=\"presentation\" aria-label=\"download icon\"><\/i>\n  Download csv file to test\n<\/a>",
>       "deps": [
>         {
>           "name": "selectize",
>           "version": "0.12.4",
>           "src": {
>             "href": "shared/selectize"
>           },
>           "meta": null,
>           "script": [
>             "js/selectize.min.js",
>             "accessibility/js/selectize-plugin-a11y.min.js"
>           ],
>           "stylesheet": "css/selectize.bootstrap3.css",
>           "head": null,
>           "attachment": null,
>           "package": null,
>           "all_files": true
>         },
>         {
>           "name": "font-awesome",
>           "version": "5.13.0",
>           "src": {
>             "href": "font-awesome-5.13.0"
>           },
>           "meta": null,
>           "script": null,
>           "stylesheet": [
>             "css/all.min.css",
>             "css/v4-shims.min.css"
>           ],
>           "head": null,
>           "attachment": null,
>           "all_files": true
>         }
>       ]
>     },
>     "ftest": null,
>     "image_down": null,
>     "note1": null,
>     "note2": null,
>     "note3": null,
>     "note4": null,
>     "note5": null,
>     "ottest": null,
>     "plot": null,
>     "pttest": null,
>     "slider": null,
>     "start_note": {
>       "html": " <h4> To perform analysis using your own dataset, prepare excel file in csv format by reading instruction below  <\/h4>\n<p>\n<ui>\n<li>Open a new blank excel file<\/li>\n<li>Copy and paste observations into a new sheet (use only one sheet) of a new excel file<\/li>\n<li>Groups to be compared should be pasted as columns <\/li>\n<li>Don't type or delete anything on other cells without data<\/li>\n<li>You can use any names for your columns. No space is allowed in the Column name. If space is required use underscore ‘_’ or ‘.’ full stop; for example ‘Variable name’ should be written as Variable_name or Variable.name<\/li>\n<li>Data should be arranged towards upper left corner and row above the data should not be left blank <\/li>\n<li>Type 'NA' in the cell with no observation<\/li>\n<li>Don't type and delete anything on other cells without data. If so select those cells, right click and click clear contents <\/li>\n<li>Give names to all column, Don't add any unnecessary columns that is not required for analysis<\/li>\n<li>Once all these are done, your file is ready. Now save it as CSV file. <\/li>\n<li><b>Upload file by clicking browse in the app <\/b><\/li>\n<\/ui>\n<\/p>\n<p>\n<b> Things to remember while doing t-test<\/b>\n<li>Note that, unpaired two-samples t-test can be used only under certain conditions: <\/li>\n<li>First condition: each of the two groups of samples being compared, should be normally distributed. This can be checked using Shapiro-Wilk test and other data visualization techniques [use grapes::descApp()]<\/li>\n<li>Second condition: Variances of the two groups should be equal.\nThis can be checked using Homogenity of variance test (F test) available\nwith in the app. <\/li>\n<li>If the variances of the two groups being compared are different (heteroscedasticity), it’s possible to use the Welch t test, an adaptation of Student t-test. <\/li>\n<li>Welch t test and F-test are available within the app<\/li>\n<\/p>\n<b> You can download a model data set from below and test the App  <\/b>\n",
>       "deps": [
>
>       ]
>     },
>     "summary1": null,
>     "summary2": null,
>     "summary3": null,
>     "summary4": null,
>     "summary5": null,
>     "ttest": null,
>     "var": null,
>     "var1": null,
>     "var2": null,
>     "var3": null,
>     "var4": null,
>     "var5": null,
>     "wttest": null
>   },
>   "export": {
>
>   }
> }

If this is expected, use `snapshotUpdate('D:/Dropbox/ILKConsultancy/grapesAgri1/inst/comp_mean')` to update
-------------------------------------------------------------------------------------------

== Results ================================================================================
Duration: 91.8 s

[ FAIL 2 | WARN 0 | SKIP 0 | PASS 0 ]
```

- [x]  Also for this line -

`Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.`

If this sounds correct change it to -

`Some of the web applications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community are user friendly but these applications don't provide options to generate plots and graphs.`

- Also not sure if we need this section here https://github.com/adithirgis/grapesAgri1/blob/master/joss/paper.md#package-dependencies-and-details-of-functions-used I will leave it to the editor.

- This is for my reference - which package did you use for styling your code?

- Also I will leave it to you the code at some places does spill over 120 characters eg-

`out = agricolae::duncan.test(d[, input$yield], d[, input$treatment], result[3, 1], result[3, 3], alpha = 0.05, group = TRUE, main = NULL, console = FALSE)`

- And the usage of `<-` is always preferred over `=` as mentioned previously. This is optional and feel free to ignore it.

Thanks!
Adithi
