### 2021-07-09 pratheesh3780/grapesAgri1#1 @pratheesh3780
> Hello,
>
> Thanks for following up soon! Really appreciate it.
>
> I have a few things in my checklist now to end the review -
>
> * [ ]  Run `devtools::check()` and `devtools::test()`, some tests seem to fail in my system. Let me add the error I get here. This is using the latest version of the app also `devtools::check()` also gives the same error.
>
**

- changed and tested again rectified the issue
- devtools::check()
-

- devtools::test()
-

**
> * [ ]   Also for this line -
>
> `Some of the webapplications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community were userfriendly and very simple to use but these applications don't provide opportunities to generate plots and graphs.`
>
> If this sounds correct change it to -
>
> `Some of the web applications like WASP (https://ccari.res.in/waspnew.html) and OPSTAT (http://14.139.232.166/opstat/) used by agricultural research community are user friendly but these applications don't provide options to generate plots and graphs.`

- **It is changed already please check**

>
> * Also not sure if we need this section here https://github.com/adithirgis/grapesAgri1/blob/master/joss/paper.md#package-dependencies-and-details-of-functions-used I will leave it to the editor.

**- ok**

> * This is for my reference - which package did you use for styling your code?

**- styler**

Thank you.
