### 2021-07-10 pratheesh3780/grapesAgri1#1 @pratheesh3780
Ok I got it; hopefully a 'loadTimeout ' in ShinyDriver$new will resolve the issue. Will get back to you shortly.

Thank you for all the support!
