### 2021-07-10 pratheesh3780/grapesAgri1#1 @pratheesh3780
Hi!
Done all what is possible....hope for the best
these are my results

==> devtools::check()

i Updating grapesAgri1 documentation
i Loading grapesAgri1
Registered S3 method overwritten by 'pryr':
  method      from
  print.bytes Rcpp
Writing NAMESPACE
Writing NAMESPACE
-- Building ---------------------------- grapesAgri1 --
Setting env vars:
* CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
* CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
* CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
-------------------------------------------------------
v  checking for file 'D:\PERMANENT ASSETS\R package\grapesAgri1/DESCRIPTION' ...
-  preparing 'grapesAgri1': (407ms)
v  checking DESCRIPTION meta-information ...
-  installing the package to process help pages
   Loading required namespace: grapesAgri1
-  saving partial Rd database (525ms)
-  checking for LF line-endings in source and make files and shell scripts (519ms)
-  checking for empty or unneeded directories
-  building 'grapesAgri1_1.0.0.tar.gz'

-- Checking ---------------------------- grapesAgri1 --
Setting env vars:
* _R_CHECK_CRAN_INCOMING_REMOTE_: FALSE
* _R_CHECK_CRAN_INCOMING_       : FALSE
* _R_CHECK_FORCE_SUGGESTS_      : FALSE
* NOT_CRAN                      : true
-- R CMD check -----------------------------------------------------------------
-  using log directory 'D:/PERMANENT ASSETS/R package/grapesAgri1.Rcheck'
-  using R version 4.0.4 (2021-02-15)
-  using platform: x86_64-w64-mingw32 (64-bit)
-  using session charset: ISO8859-1
-  using options '--no-manual --as-cran'
v  checking for file 'grapesAgri1/DESCRIPTION'
-  checking extension type ... Package
-  this is package 'grapesAgri1' version '1.0.0'
-  package encoding: UTF-8
v  checking package namespace information
v  checking package dependencies (5.4s)
v  checking if this is a source package ...
v  checking if there is a namespace
v  checking for executable files (689ms)
v  checking for hidden files and directories ...
v  checking for portable file names ...
v  checking serialization versions ...
v  checking whether package 'grapesAgri1' can be installed (3.3s)
v  checking installed package size ...
v  checking package directory (695ms)
v  checking for future file timestamps (32.4s)
v  checking DESCRIPTION meta-information (352ms)
v  checking top-level files
v  checking for left-over files ...
v  checking index information
v  checking package subdirectories ...
v  checking R files for non-ASCII characters ...
v  checking R files for syntax errors ...
v  checking whether the package can be loaded ...
v  checking whether the package can be loaded with stated dependencies ...
v  checking whether the package can be unloaded cleanly ...
v  checking whether the namespace can be loaded with stated dependencies ...
v  checking whether the namespace can be unloaded cleanly (441ms)
v  checking loading without being on the library search path (765ms)
v  checking dependencies in R code (561ms)
v  checking S3 generic/method consistency (550ms)
v  checking replacement functions (344ms)
v  checking foreign function calls (337ms)
v  checking R code for possible problems (2.5s)
v  checking Rd files (893ms)
v  checking Rd metadata ...
v  checking Rd line widths ...
v  checking Rd cross-references ...
v  checking for missing documentation entries (340ms)
v  checking for code/documentation mismatches (796ms)
v  checking Rd \usage sections (782ms)
v  checking Rd contents ...
v  checking for unstated dependencies in examples ...
v  checking examples (660ms)
v  checking for unstated dependencies in 'tests' ...
-  checking tests ...
v  Running 'testthat.R' [20s] (19.9s)
v  checking for non-standard things in the check directory (19.9s)
v  checking for detritus in the temp directory

-- R CMD check results ---------------------------------- grapesAgri1 1.0.0 ----
Duration: 1m 15.8s

0 errors v | 0 warnings v | 0 notes v

R CMD check succeeded

--------------------------------------------------------------------------------------------------

==> devtools::test()

i Loading grapesAgri1
Registered S3 method overwritten by 'pryr':
  method      from
  print.bytes Rcpp
i Testing grapesAgri1
v |  OK F W S | Context
v |   1   11   | app-file [11.7 s]
-------------------------------------------------------
Warning (testdesc.R:2:1): (code run outside of `test_that()`)
package 'shinytest' was built under R version 4.0.5
Backtrace:
 1. base::library(shinytest) testdesc.R:2:0
 2. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'shiny' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny::loadSupport(appDir, renv = sharedEnv, globalrenv = NULL)
 17. base::require(shiny)
 22. base::library(...)
 23. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'shinyWidgets' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(shinyWidgets)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'rmarkdown' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(rmarkdown)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'knitr' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(knitr)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'dplyr' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(dplyr)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'gridGraphics' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(gridGraphics)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'kableExtra' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(kableExtra)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'pastecs' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(pastecs)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'ggplot2' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(ggpubr)
 21. base::.getRequiredPackages2(pkgInfo, quietly = quietly)
 22. base::library(...)
 23. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testdesc.R:6:3): descApp_testing
package 'summarytools' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testdesc.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(summarytools)
 21. base:::testRversion(pkgInfo, package, pkgpath)
-------------------------------------------------------
v |   1   3   | app-file [6.1 s]
-------------------------------------------------------
Warning (testttest.R:6:3): ttest_testing
package 'PairedData' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testttest.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(PairedData)
 21. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testttest.R:6:3): ttest_testing
package 'gld' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testttest.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(PairedData)
 21. base::.getRequiredPackages2(pkgInfo, quietly = quietly)
 22. base::library(...)
 23. base:::testRversion(pkgInfo, package, pkgpath)

Warning (testttest.R:6:3): ttest_testing
package 'reshape2' was built under R version 4.0.5
Backtrace:
  1. shinytest::expect_pass(testApp(appDir, compareImages = FALSE)) testttest.R:6:2
  2. shinytest::testApp(appDir, compareImages = FALSE)
  3. shinytest:::app_path(appDir, "appDir")
  4. shinytest:::is_app(path)
 12. shiny::shinyAppDir(path)
 13. shiny:::shinyAppDir_appR("app.R", appDir, options = options)
 14. shiny:::appObj()
 15. shiny:::func(fname, ...)
 16. shiny:::sourceUTF8(fullpath, envir = new.env(parent = sharedEnv))
 17. [ base::eval(...) ] with 1 more call
 20. base::library(reshape2)
 21. base:::testRversion(pkgInfo, package, pkgpath)
-------------------------------------------------------

== Results ============================================
Duration: 17.9 s

[ FAIL 0 | WARN 14 | SKIP 0 | PASS 2 ]

Nice code.
