+++ .github/workflows/R-CMD-check.yaml
@@ -64,6 +64,14 @@ jobs:
           sysreqs=$(Rscript -e "cat(sysreqs::sysreq_commands('DESCRIPTION'))")
           sudo -s eval "$sysreqs"
 
+      - name: Install xquartz
+        if: runner.os == 'macOS'
+        run: brew install --cask xquartz
+
+      - name: Install xquartz
+        if: runner.os == 'windows'
+        run: brew install --cask xquartz
+
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
