### 2021-07-11 pratheesh3780/grapesAgri1#1 @adithirgis
Hello,

I still get the error, I think its screen shot issue which can arises if the size of the screens differ (this is an issue with shinytest package). I will wrap up the review now. Congratulations!

And three optional comments -

- styling of the code can be restricted to 120 characters
- usage of `<-` instead of `=` is preferred
- and reducing the number of packages imported, getting the package into CRAN will be difficult, suggestion would be to use base R packages.

Thanks!
Adithi
