+++ inst/desc/tests/shinytest/mytest.R
@@ -1,20 +1,4 @@
-app <- ShinyDriver$new("../../", loadTimeout = 1e+05, seed =1234 )
+app <- ShinyDriver$new("../../")
 app$snapshotInit("mytest")
 
 app$snapshot()
-app$uploadFile(file1 = "iris.csv") # <-- This should be the path to the file, relative to the app's tests/shinytest directory
-app$setInputs(var = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(submit4 = "click",allowInputNoBinding_ = TRUE)
-app$setInputs(req = "sumbygrp")
-app$setInputs(var = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(group = "Species",allowInputNoBinding_ = TRUE)
-app$setInputs(submit5 = "click",allowInputNoBinding_ = TRUE)
-app$setInputs(req = "boxplot",allowInputNoBinding_ = TRUE)
-app$setInputs(variable = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),allowInputNoBinding_ = TRUE)
-app$setInputs(submit1 = "click",allowInputNoBinding_ = TRUE)
-app$setInputs(req = "histogram",allowInputNoBinding_ = TRUE)
-app$setInputs(submit2 = "click",allowInputNoBinding_ = TRUE)
-app$setInputs(req = "qqplot",allowInputNoBinding_ = TRUE)
-app$setInputs(submit3 = "click",allowInputNoBinding_ = TRUE)
-app$setInputs(req = "nt",allowInputNoBinding_ = TRUE)
-app$setInputs(submit = "click",allowInputNoBinding_ = TRUE)
