+++ inst/comp_mean/tests/shinytest/mytest.R
@@ -1,14 +1,4 @@
-app <- ShinyDriver$new("../../", loadTimeout = 1e+05, seed =1234)
+app <- ShinyDriver$new("../../")
 app$snapshotInit("mytest")
 
 app$snapshot()
-app$uploadFile(file1 = "iris.csv")
-app$setInputs(submit1 = "click")
-app$setInputs(req1 = "pttest")
-app$setInputs(submit5 = "click")
-app$setInputs(req1 = "ottest")
-app$setInputs(submit3 = "click")
-app$setInputs(req1 = "wttest")
-app$setInputs(submit2 = "click")
-app$setInputs(req1 = "ftest")
-app$setInputs(submit4 = "click")
