+++ tests/testthat/testttest.R
@@ -1,7 +1,0 @@
-context("app-file")
-library(shinytest)
-test_that("ttest_testing", {
-  skip_on_cran()
-  appDir <- system.file("comp_mean", package = "grapesAgri1")
-  expect_pass(testApp(appDir, compareImages = FALSE))
-})
