+++ inst/desc/tests/shinytest/mytest.R
@@ -1,4 +1,0 @@
-app <- ShinyDriver$new("../../", loadTimeout= 1e+05, seed=1234)
-app$snapshotInit("mytest")
-
-app$snapshot()
