+++ inst/crd/tests/shinytest/mytest.R
@@ -3,6 +3,6 @@ app$snapshotInit("mytest")
 
 app$snapshot()
 app$uploadFile(file1 = "equal_replication.csv")
-app$setInputs(trt = 3,allowInputNoBinding_ = TRUE)
-app$setInputs(yield = "WEIGHT",allowInputNoBinding_ = TRUE)
+app$setInputs(trt = 3)
+app$setInputs(yield = "WEIGHT")
 app$setInputs(submit = "click")
