+++ inst/RBD/app.R
@@ -3,7 +3,6 @@ library(ggplot2)
 library(Hmisc)
 library(agricolae)
 library(shinyWidgets)
-library(shinycssloaders)
 library(dplyr)
 library(rmarkdown)
 library(RColorBrewer)
