+++ inst/RBD/tests/shinytest/mytest.R
@@ -1,10 +1,0 @@
-app <- ShinyDriver$new("../../")
-app$snapshotInit("mytest")
-
-app$snapshot()
-app$uploadFile(file1 = "mango_data.csv")
-app$setInputs(trt = 10)
-app$setInputs(rep = 3)
-app$setInputs(Replication = "Replication")
-app$setInputs(yield = "yield")
-app$setInputs(submit = "click")
