+++ tests/testthat/testdesc.R
@@ -1,7 +1,7 @@
 context("app-file")
 library(shinytest)
-test_that("rbdApp_testing", {
+test_that("descApp_testing", {
   skip_on_cran()
-  appDir <- system.file("RBD", package = "grapesAgri1")
+  appDir <- system.file("desc", package = "grapesAgri1")
   expect_pass(testApp(appDir, compareImages = FALSE))
 })
