+++ inst/layout/tests/shinytest/mytest.R
@@ -1,7 +1,4 @@
-app <- ShinyDriver$new("../../")
+app <- ShinyDriver$new("../../", loadTimeout = 1e+05)
 app$snapshotInit("mytest")
 
 app$snapshot()
-app$setInputs(trt_crd = 3)
-app$setInputs(rep_crd = 5)
-app$setInputs(submit1 = "click")
