+++ inst/Corr/tests/shinytest/mytest.R
@@ -1,5 +1,5 @@
 app <- ShinyDriver$new("../../", loadTimeout = 1e+05)
 app$snapshotInit("mytest")
 
-app$uploadFile(file1 = "iris.csv") # <-- This should be the path to the file, relative to the app's tests/shinytest directory
 app$snapshot()
+
