### 2021-07-11 pratheesh3780/grapesAgri1#1 @pratheesh3780
Hi,
I also noticed that, I was working on that ....R-CMD check is passing for mac and ubuntu but not for windows. Trying to resolve.
let me also add ....really appreciate your efforts in supporting open-source initiatives. Lucky to have you as the reviewer. Learnt a lot!
Thank you!
