### 2021-07-11 pratheesh3780/grapesAgri1#1 @pratheesh3780
Hi,
finally resolved the issue, R-CMD check is passing. It might be related to how git on Windows converts \n to \r\n. so resolved it by adding
'- name: Windows git setup
        if: runner.os == 'Windows'
        run:
          git config --global core.autocrlf false'
in yaml

phew!! 😌

I think issue is I should keep my git config core.autocrlf false, so you can clone in to a windows system without any issues.

Thank You
