### 2021-07-11 pratheesh3780/grapesAgri1#1 @pratheesh3780
Hi
You can also try set autocrlf to false before pull
git config --global core.autocrlf false
and please let me know if this works,then I can add rest of the tests.
Thank you!
