### 2021-07-12 pratheesh3780/grapesAgri1#1 @adithirgis
Hi,

I checked again but I still get the errors, but I am pretty sure its because the sizes of the image issue. I think its not resolvable as of now using shinytest. Check this - https://github.com/rstudio/shinytest/issues/300 and https://github.com/rstudio/shinytest/issues/315.

Thanks
