### 2021-07-13 pratheesh3780/grapesAgri1#3 @elimillera
JOSS Review: Raw HTML is found in page titles
Hello 👋

The use of tags in `titlePanel` results in raw html being displayed in the page title. Using the `windowTitle` argument will override this behavior.

![image](https://user-images.githubusercontent.com/24767886/125377299-5eaab000-e352-11eb-9d73-059bf2ae040d.png)
