### 2021-07-13 pratheesh3780/grapesAgri1#5 @elimillera
JOSS Review: Observe events for submit buttons
This is related to #4 . Not necessarily something that should hold up a review, just a recommendation.

In several places `input$submit > 0` or `is.null(input$submit)` is used. If you are only looking to update the displays when a submitting button is pressed, the `shiny::observeEvent` function would be the solution to that. You could wrap the derivation of tables and plots in that to only have the outputs change on a update to `input$submit`. Right now the application is reactive when the submit button has been pressed but not otherwise.
