### 2021-07-13 #3437 @elimillera
@adithirgis @fabian-s
My review is done.

Both https://github.com/pratheesh3780/grapesAgri1/issues/4 and https://github.com/pratheesh3780/grapesAgri1/issues/3 are small issues that don't effect functionality. I also have some small comments in https://github.com/pratheesh3780/grapesAgri1/pull/6.

 I would think https://github.com/pratheesh3780/grapesAgri1/issues/2 is the only issue that I would want resolved before release.

 Thank you for the invitation and your work on the application!
