+++ inst/layout/app.R
@@ -8,13 +8,16 @@ library(desplot)
 library(knitr)
 
 ############################## Ui
-ui = fluidPage(
+ui <- fluidPage(
   setBackgroundColor(
     color = c("#fff291", "#ffffff"),
     gradient = "radial",
     direction = c("bottom", "right")
   ),
-  titlePanel(tags$div(tags$b("Field Layout of experiments"), style = "color:#000000")),
+  titlePanel(
+    title = tags$div(tags$b("Field Layout of experiments"), style = "color:#000000"),
+    windowTitle = "Field Layout"
+  ),
   sidebarPanel(
     selectInput(
       "design", tags$div(tags$b("Select the design for your experiment"),
@@ -151,10 +154,10 @@ ui = fluidPage(
   )
 )
 
-server = function(input, output, session) {
+server <- function(input, output, session) {
 
   ################### CRD
-  output$var_crd = renderUI({
+  output$var_crd <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -163,7 +166,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$text_crd1 = renderUI({
+  output$text_crd1 <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -172,8 +175,8 @@ server = function(input, output, session) {
     }
     if (input$design == "crd") {
       if (input$trt_crd >= 2) {
-        t = input$trt_crd
-        minrep.req = ((12 + t) / t)
+        t <- input$trt_crd
+        minrep.req <- ((12 + t) / t)
         HTML(paste0(tags$b(" minimum number of replication
                             required is:", ceiling(minrep.req))))
       }
@@ -183,7 +186,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$var_crd1 = renderUI({
+  output$var_crd1 <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -215,7 +218,7 @@ server = function(input, output, session) {
   })
 
 
-  output$details_crd = renderTable(
+  output$details_crd <- renderTable(
     {
       if (is.null(input$rep_crd)) {
         return()
@@ -231,11 +234,11 @@ server = function(input, output, session) {
       }
       if (input$design == "crd") {
         if (input$rep_crd > 0 && input$submit1 > 0) {
-          Treatments = input$trt_crd
-          Replication = input$rep_crd
-          Experimental.units = Treatments * Replication
-          det = cbind(Treatments, Replication, Experimental.units)
-          det = as.data.frame(det)
+          Treatments <- input$trt_crd
+          Replication <- input$rep_crd
+          Experimental.units <- Treatments * Replication
+          det <- cbind(Treatments, Replication, Experimental.units)
+          det <- as.data.frame(det)
           det
         }
         else {
@@ -250,7 +253,7 @@ server = function(input, output, session) {
     rownames = FALSE
   )
 
-  output$text_crd2 = renderUI({
+  output$text_crd2 <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -271,7 +274,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$text_crd3 = renderUI({
+  output$text_crd3 <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -298,7 +301,7 @@ server = function(input, output, session) {
   ############################# END CRD
 
   ################### RBD
-  output$var_rbd = renderUI({
+  output$var_rbd <- renderUI({
     if (is.null(input$design)) {
       return()
     }
@@ -307,7 +310,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$text_rbd1 = renderUI({
+  output$text_rbd1 <- renderUI({
[…]
