+++ inst/RBD/app.R
@@ -8,13 +8,16 @@ library(rmarkdown)
 library(RColorBrewer)
 library(ggpubr)
 
-ui = fluidPage(
+ui <- fluidPage(
   setBackgroundColor(
     color = c("#ffb9b3", "#ffffff"),
     gradient = "radial",
     direction = c("bottom", "right")
   ),
-  titlePanel(tags$div(tags$b("Two-way Analysis of Variance", style = "color:#000000"))),
+  titlePanel(
+    title = tags$div(tags$b("Two-way Analysis of Variance", style = "color:#000000")),
+    windowTitle = "Randomized Block Design"
+  ),
   sidebarPanel(
     fileInput("file1", "CSV File (upload in csv format)", accept = c("text/csv", "text/comma-separated-values,text/plain", ".csv")),
     checkboxInput("header", "Header", TRUE),
@@ -106,17 +109,17 @@ ui = fluidPage(
   )
 )
 ####################
-server = function(input, output, session) {
-  csvfile = reactive({
-    csvfile = input$file1
+server <- function(input, output, session) {
+  csvfile <- reactive({
+    csvfile <- input$file1
     if (is.null(csvfile)) {
       return(NULL)
     }
-    dt = read.csv(csvfile$datapath, header = input$header, sep = ",")
+    dt <- read.csv(csvfile$datapath, header = input$header, sep = ",")
     dt
   })
 
-  output$var = renderUI({
+  output$var <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -138,7 +141,7 @@ server = function(input, output, session) {
     }
   })
 
-  output$var1 = renderUI({
+  output$var1 <- renderUI({
     if (is.null(input$file1$datapath)) {
       return()
     }
@@ -156,7 +159,7 @@ server = function(input, output, session) {
   })
 
   ##################################### TRT means
-  output$trtmeans = renderTable(
+  output$trtmeans <- renderTable(
     {
       if (is.null(input$file1$datapath)) {
         return()
@@ -182,20 +185,20 @@ server = function(input, output, session) {
         )
         input$reload
         Sys.sleep(2)
-        d = as.data.frame(csvfile())
-        r = as.numeric(input$rep)
-        t = as.numeric(input$trt)
-        response = d[, input$yield]
-        treatment = d[, input$treatment]
-        replication = d[, input$Replication]
-        treatment = factor(treatment)
-        replication = factor(replication)
-        anvaTable = lm(response ~ treatment + replication)
-        result = as.data.frame(anova(anvaTable))
-        out = agricolae::LSD.test(csvfile()[, input$yield], csvfile()[, input$treatment], result[3, 1], result[3, 3])
-        trtmeans = out$means
-        colnames(trtmeans)[1] = "Treatment_means"
-        drops = c("r", "Q25", "Q50", "Q75")
+        d <- as.data.frame(csvfile())
+        r <- as.numeric(input$rep)
+        t <- as.numeric(input$trt)
+        response <- d[, input$yield]
+        treatment <- d[, input$treatment]
+        replication <- d[, input$Replication]
+        treatment <- factor(treatment)
+        replication <- factor(replication)
+        anvaTable <- lm(response ~ treatment + replication)
+        result <- as.data.frame(anova(anvaTable))
+        out <- agricolae::LSD.test(csvfile()[, input$yield], csvfile()[, input$treatment], result[3, 1], result[3, 3])
+        trtmeans <- out$means
+        colnames(trtmeans)[1] <- "Treatment_means"
+        drops <- c("r", "Q25", "Q50", "Q75")
         trtmeans[, !(names(trtmeans) %in% drops)]
       }
     },
[…]
