### 2021-07-14 5d5d034
proofreading edits
almost good to go here -- adjusted capitalization and formatted shiny like the other R packages you mention in the text.
