### 2021-07-14 #3437 @fabian-s
@pratheesh3780

1. do you want to work on the coding issues (https://github.com/pratheesh3780/grapesAgri1/issues/5, https://github.com/pratheesh3780/grapesAgri1/issues/4) raised by @elimillera before we define the JOSS release of your package? Not required, up to you. (The JOSS paper will point to the version of your package that was available at publication, so it might be a good idea to improve the code first..)
2. please see my proofreading edits at https://github.com/pratheesh3780/grapesAgri1/pull/7 and merge the ones that seem reasonable to you
3. the "missing DOI" whedon complains about is a false positive, just ignore that.
