### 2021-07-14 #3437 @pratheesh3780
@fabian-s
1. As the app is stable and working fine, I don't intend to make those changes. Though it is a valuable suggestion.
2. Proofreading edits were merged
3. ok

This was my first experience  with JOSS. Such a developer friendly platform. Future of scientific journals is here. I enjoyed the entire review process and valuable suggestions from @adithirgis  and @elimillera . Learned a lot.

Thank you!
