### 2021-07-15 #3437 @pratheesh3780
@fabian-s
Please see the Release details
Version 1.1.0
DOI: 10.5281/zenodo.5106216

Thank you!
