### 2021-07-16 #3437 @fabian-s
@pratheesh3780 :

seems like your Zenodo archive doesn't have correct metadata yet:

Authors are given as  "pratheesh3780; Eli Miller; Fabian Scheipl; Adithi R. Upadhya",
should be "Pratheesh P. Gopinath, Rajender Parsad, Brigit Joseph, Adarsh V. S." (same as your JOSS paper)

Title is "pratheesh3780/grapesAgri1: grapesAgri1- Collection of R-shiny Package for Agricultural Data Analysis Part 1"
should be "grapesAgri1- Collection of R-shiny Package for Agricultural Data Analysis" (same as your JOSS paper)

Please fix ASAP and let me know here once that's done.
